% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbreviate_signature_linter.R
\name{abbreviate_signature_linter}
\alias{abbreviate_signature_linter}
\title{Check signature of \code{\link[=abbreviate]{abbreviate()}}}
\usage{
abbreviate_signature_linter()
}
\description{
Check signature of \code{\link[=abbreviate]{abbreviate()}}
}
