% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_signature_linter.R
\name{abs_signature_linter}
\alias{abs_signature_linter}
\title{Check signature of \code{\link[=abs]{abs()}}}
\usage{
abs_signature_linter()
}
\description{
Check signature of \code{\link[=abs]{abs()}}
}
