% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acos_signature_linter.R
\name{acos_signature_linter}
\alias{acos_signature_linter}
\title{Check signature of \code{\link[=acos]{acos()}}}
\usage{
acos_signature_linter()
}
\description{
Check signature of \code{\link[=acos]{acos()}}
}
