% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acosh_signature_linter.R
\name{acosh_signature_linter}
\alias{acosh_signature_linter}
\title{Check signature of \code{\link[=acosh]{acosh()}}}
\usage{
acosh_signature_linter()
}
\description{
Check signature of \code{\link[=acosh]{acosh()}}
}
