% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activeBindingFunction_signature_linter.R
\name{activeBindingFunction_signature_linter}
\alias{activeBindingFunction_signature_linter}
\title{Check signature of \code{\link[=activeBindingFunction]{activeBindingFunction()}}}
\usage{
activeBindingFunction_signature_linter()
}
\description{
Check signature of \code{\link[=activeBindingFunction]{activeBindingFunction()}}
}
