% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNA_signature_linter.R
\name{addNA_signature_linter}
\alias{addNA_signature_linter}
\title{Check signature of \code{\link[=addNA]{addNA()}}}
\usage{
addNA_signature_linter()
}
\description{
Check signature of \code{\link[=addNA]{addNA()}}
}
