% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTaskCallback_signature_linter.R
\name{addTaskCallback_signature_linter}
\alias{addTaskCallback_signature_linter}
\title{Check signature of \code{\link[=addTaskCallback]{addTaskCallback()}}}
\usage{
addTaskCallback_signature_linter()
}
\description{
Check signature of \code{\link[=addTaskCallback]{addTaskCallback()}}
}
