% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrep_signature_linter.R
\name{agrep_signature_linter}
\alias{agrep_signature_linter}
\title{Check signature of \code{\link[=agrep]{agrep()}}}
\usage{
agrep_signature_linter()
}
\description{
Check signature of \code{\link[=agrep]{agrep()}}
}
