% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrepl_signature_linter.R
\name{agrepl_signature_linter}
\alias{agrepl_signature_linter}
\title{Check signature of \code{\link[=agrepl]{agrepl()}}}
\usage{
agrepl_signature_linter()
}
\description{
Check signature of \code{\link[=agrepl]{agrepl()}}
}
