% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_equal_formula_signature_linter.R
\name{all_equal_formula_signature_linter}
\alias{all_equal_formula_signature_linter}
\title{Check signature of \code{\link[=all.equal.formula]{all.equal.formula()}}}
\usage{
all_equal_formula_signature_linter()
}
\description{
Check signature of \code{\link[=all.equal.formula]{all.equal.formula()}}
}
