% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_names_signature_linter.R
\name{all_names_signature_linter}
\alias{all_names_signature_linter}
\title{Check signature of \code{\link[=all.names]{all.names()}}}
\usage{
all_names_signature_linter()
}
\description{
Check signature of \code{\link[=all.names]{all.names()}}
}
