% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_vars_signature_linter.R
\name{all_vars_signature_linter}
\alias{all_vars_signature_linter}
\title{Check signature of \code{\link[=all.vars]{all.vars()}}}
\usage{
all_vars_signature_linter()
}
\description{
Check signature of \code{\link[=all.vars]{all.vars()}}
}
