% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allowInterrupts_signature_linter.R
\name{allowInterrupts_signature_linter}
\alias{allowInterrupts_signature_linter}
\title{Check signature of \code{\link[=allowInterrupts]{allowInterrupts()}}}
\usage{
allowInterrupts_signature_linter()
}
\description{
Check signature of \code{\link[=allowInterrupts]{allowInterrupts()}}
}
