% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyDuplicated_array_signature_linter.R
\name{anyDuplicated_array_signature_linter}
\alias{anyDuplicated_array_signature_linter}
\title{Check signature of \code{\link[=anyDuplicated.array]{anyDuplicated.array()}}}
\usage{
anyDuplicated_array_signature_linter()
}
\description{
Check signature of \code{\link[=anyDuplicated.array]{anyDuplicated.array()}}
}
