% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyDuplicated_matrix_signature_linter.R
\name{anyDuplicated_matrix_signature_linter}
\alias{anyDuplicated_matrix_signature_linter}
\title{Check signature of \code{\link[=anyDuplicated.matrix]{anyDuplicated.matrix()}}}
\usage{
anyDuplicated_matrix_signature_linter()
}
\description{
Check signature of \code{\link[=anyDuplicated.matrix]{anyDuplicated.matrix()}}
}
