% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyNA_data_frame_signature_linter.R
\name{anyNA_data_frame_signature_linter}
\alias{anyNA_data_frame_signature_linter}
\title{Check signature of \code{\link[=anyNA.data.frame]{anyNA.data.frame()}}}
\usage{
anyNA_data_frame_signature_linter()
}
\description{
Check signature of \code{\link[=anyNA.data.frame]{anyNA.data.frame()}}
}
