% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyNA_numeric_version_signature_linter.R
\name{anyNA_numeric_version_signature_linter}
\alias{anyNA_numeric_version_signature_linter}
\title{Check signature of \code{\link[=anyNA.numeric_version]{anyNA.numeric_version()}}}
\usage{
anyNA_numeric_version_signature_linter()
}
\description{
Check signature of \code{\link[=anyNA.numeric_version]{anyNA.numeric_version()}}
}
