% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyNA_POSIXlt_signature_linter.R
\name{anyNA_POSIXlt_signature_linter}
\alias{anyNA_POSIXlt_signature_linter}
\title{Check signature of \code{\link[=anyNA.POSIXlt]{anyNA.POSIXlt()}}}
\usage{
anyNA_POSIXlt_signature_linter()
}
\description{
Check signature of \code{\link[=anyNA.POSIXlt]{anyNA.POSIXlt()}}
}
