% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyNA_signature_linter.R
\name{anyNA_signature_linter}
\alias{anyNA_signature_linter}
\title{Check signature of \code{\link[=anyNA]{anyNA()}}}
\usage{
anyNA_signature_linter()
}
\description{
Check signature of \code{\link[=anyNA]{anyNA()}}
}
