% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aperm_table_signature_linter.R
\name{aperm_table_signature_linter}
\alias{aperm_table_signature_linter}
\title{Check signature of \code{\link[=aperm.table]{aperm.table()}}}
\usage{
aperm_table_signature_linter()
}
\description{
Check signature of \code{\link[=aperm.table]{aperm.table()}}
}
