% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_signature_linter.R
\name{append_signature_linter}
\alias{append_signature_linter}
\title{Check signature of \code{\link[=append]{append()}}}
\usage{
append_signature_linter()
}
\description{
Check signature of \code{\link[=append]{append()}}
}
