% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arg_signature_linter.R
\name{Arg_signature_linter}
\alias{Arg_signature_linter}
\title{Check signature of \code{\link[=Arg]{Arg()}}}
\usage{
Arg_signature_linter()
}
\description{
Check signature of \code{\link[=Arg]{Arg()}}
}
