% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args_signature_linter.R
\name{args_signature_linter}
\alias{args_signature_linter}
\title{Check signature of \code{\link[=args]{args()}}}
\usage{
args_signature_linter()
}
\description{
Check signature of \code{\link[=args]{args()}}
}
