% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_signature_linter.R
\name{array_signature_linter}
\alias{array_signature_linter}
\title{Check signature of \code{\link[=array]{array()}}}
\usage{
array_signature_linter()
}
\description{
Check signature of \code{\link[=array]{array()}}
}
