% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayInd_signature_linter.R
\name{arrayInd_signature_linter}
\alias{arrayInd_signature_linter}
\title{Check signature of \code{\link[=arrayInd]{arrayInd()}}}
\usage{
arrayInd_signature_linter()
}
\description{
Check signature of \code{\link[=arrayInd]{arrayInd()}}
}
