% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_array_default_signature_linter.R
\name{as_array_default_signature_linter}
\alias{as_array_default_signature_linter}
\title{Check signature of \code{\link[=as.array.default]{as.array.default()}}}
\usage{
as_array_default_signature_linter()
}
\description{
Check signature of \code{\link[=as.array.default]{as.array.default()}}
}
