% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_call_signature_linter.R
\name{as_call_signature_linter}
\alias{as_call_signature_linter}
\title{Check signature of \code{\link[=as.call]{as.call()}}}
\usage{
as_call_signature_linter()
}
\description{
Check signature of \code{\link[=as.call]{as.call()}}
}
