% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_character_error_signature_linter.R
\name{as_character_error_signature_linter}
\alias{as_character_error_signature_linter}
\title{Check signature of \code{\link[=as.character.error]{as.character.error()}}}
\usage{
as_character_error_signature_linter()
}
\description{
Check signature of \code{\link[=as.character.error]{as.character.error()}}
}
