% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_character_octmode_signature_linter.R
\name{as_character_octmode_signature_linter}
\alias{as_character_octmode_signature_linter}
\title{Check signature of \code{\link[=as.character.octmode]{as.character.octmode()}}}
\usage{
as_character_octmode_signature_linter()
}
\description{
Check signature of \code{\link[=as.character.octmode]{as.character.octmode()}}
}
