% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame_AsIs_signature_linter.R
\name{as_data_frame_AsIs_signature_linter}
\alias{as_data_frame_AsIs_signature_linter}
\title{Check signature of \code{\link[=as.data.frame.AsIs]{as.data.frame.AsIs()}}}
\usage{
as_data_frame_AsIs_signature_linter()
}
\description{
Check signature of \code{\link[=as.data.frame.AsIs]{as.data.frame.AsIs()}}
}
