% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame_integer_signature_linter.R
\name{as_data_frame_integer_signature_linter}
\alias{as_data_frame_integer_signature_linter}
\title{Check signature of \code{\link[=as.data.frame.integer]{as.data.frame.integer()}}}
\usage{
as_data_frame_integer_signature_linter()
}
\description{
Check signature of \code{\link[=as.data.frame.integer]{as.data.frame.integer()}}
}
