% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame_ts_signature_linter.R
\name{as_data_frame_ts_signature_linter}
\alias{as_data_frame_ts_signature_linter}
\title{Check signature of \code{\link[=as.data.frame.ts]{as.data.frame.ts()}}}
\usage{
as_data_frame_ts_signature_linter()
}
\description{
Check signature of \code{\link[=as.data.frame.ts]{as.data.frame.ts()}}
}
