% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_Date_default_signature_linter.R
\name{as_Date_default_signature_linter}
\alias{as_Date_default_signature_linter}
\title{Check signature of \code{\link[=as.Date.default]{as.Date.default()}}}
\usage{
as_Date_default_signature_linter()
}
\description{
Check signature of \code{\link[=as.Date.default]{as.Date.default()}}
}
