% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_difftime_signature_linter.R
\name{as_difftime_signature_linter}
\alias{as_difftime_signature_linter}
\title{Check signature of \code{\link[=as.difftime]{as.difftime()}}}
\usage{
as_difftime_signature_linter()
}
\description{
Check signature of \code{\link[=as.difftime]{as.difftime()}}
}
