% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_double_difftime_signature_linter.R
\name{as_double_difftime_signature_linter}
\alias{as_double_difftime_signature_linter}
\title{Check signature of \code{\link[=as.double.difftime]{as.double.difftime()}}}
\usage{
as_double_difftime_signature_linter()
}
\description{
Check signature of \code{\link[=as.double.difftime]{as.double.difftime()}}
}
