% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_environment_signature_linter.R
\name{as_environment_signature_linter}
\alias{as_environment_signature_linter}
\title{Check signature of \code{\link[=as.environment]{as.environment()}}}
\usage{
as_environment_signature_linter()
}
\description{
Check signature of \code{\link[=as.environment]{as.environment()}}
}
