% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_factor_signature_linter.R
\name{as_factor_signature_linter}
\alias{as_factor_signature_linter}
\title{Check signature of \code{\link[=as.factor]{as.factor()}}}
\usage{
as_factor_signature_linter()
}
\description{
Check signature of \code{\link[=as.factor]{as.factor()}}
}
