% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_hexmode_signature_linter.R
\name{as_hexmode_signature_linter}
\alias{as_hexmode_signature_linter}
\title{Check signature of \code{\link[=as.hexmode]{as.hexmode()}}}
\usage{
as_hexmode_signature_linter()
}
\description{
Check signature of \code{\link[=as.hexmode]{as.hexmode()}}
}
