% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_list_data_frame_signature_linter.R
\name{as_list_data_frame_signature_linter}
\alias{as_list_data_frame_signature_linter}
\title{Check signature of \code{\link[=as.list.data.frame]{as.list.data.frame()}}}
\usage{
as_list_data_frame_signature_linter()
}
\description{
Check signature of \code{\link[=as.list.data.frame]{as.list.data.frame()}}
}
