% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_list_function_signature_linter.R
\name{as_list_function_signature_linter}
\alias{as_list_function_signature_linter}
\title{Check signature of \code{\link[=as.list.function]{as.list.function()}}}
\usage{
as_list_function_signature_linter()
}
\description{
Check signature of \code{\link[=as.list.function]{as.list.function()}}
}
