% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_list_numeric_version_signature_linter.R
\name{as_list_numeric_version_signature_linter}
\alias{as_list_numeric_version_signature_linter}
\title{Check signature of \code{\link[=as.list.numeric_version]{as.list.numeric_version()}}}
\usage{
as_list_numeric_version_signature_linter()
}
\description{
Check signature of \code{\link[=as.list.numeric_version]{as.list.numeric_version()}}
}
