% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_logical_factor_signature_linter.R
\name{as_logical_factor_signature_linter}
\alias{as_logical_factor_signature_linter}
\title{Check signature of \code{\link[=as.logical.factor]{as.logical.factor()}}}
\usage{
as_logical_factor_signature_linter()
}
\description{
Check signature of \code{\link[=as.logical.factor]{as.logical.factor()}}
}
