% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_matrix_noquote_signature_linter.R
\name{as_matrix_noquote_signature_linter}
\alias{as_matrix_noquote_signature_linter}
\title{Check signature of \code{\link[=as.matrix.noquote]{as.matrix.noquote()}}}
\usage{
as_matrix_noquote_signature_linter()
}
\description{
Check signature of \code{\link[=as.matrix.noquote]{as.matrix.noquote()}}
}
