% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_name_signature_linter.R
\name{as_name_signature_linter}
\alias{as_name_signature_linter}
\title{Check signature of \code{\link[=as.name]{as.name()}}}
\usage{
as_name_signature_linter()
}
\description{
Check signature of \code{\link[=as.name]{as.name()}}
}
