% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric_version_signature_linter.R
\name{as_numeric_version_signature_linter}
\alias{as_numeric_version_signature_linter}
\title{Check signature of \code{\link[=as.numeric_version]{as.numeric_version()}}}
\usage{
as_numeric_version_signature_linter()
}
\description{
Check signature of \code{\link[=as.numeric_version]{as.numeric_version()}}
}
