% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_octmode_signature_linter.R
\name{as_octmode_signature_linter}
\alias{as_octmode_signature_linter}
\title{Check signature of \code{\link[=as.octmode]{as.octmode()}}}
\usage{
as_octmode_signature_linter()
}
\description{
Check signature of \code{\link[=as.octmode]{as.octmode()}}
}
