% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_ordered_signature_linter.R
\name{as_ordered_signature_linter}
\alias{as_ordered_signature_linter}
\title{Check signature of \code{\link[=as.ordered]{as.ordered()}}}
\usage{
as_ordered_signature_linter()
}
\description{
Check signature of \code{\link[=as.ordered]{as.ordered()}}
}
