% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_package_version_signature_linter.R
\name{as_package_version_signature_linter}
\alias{as_package_version_signature_linter}
\title{Check signature of \code{\link[=as.package_version]{as.package_version()}}}
\usage{
as_package_version_signature_linter()
}
\description{
Check signature of \code{\link[=as.package_version]{as.package_version()}}
}
