% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_pairlist_signature_linter.R
\name{as_pairlist_signature_linter}
\alias{as_pairlist_signature_linter}
\title{Check signature of \code{\link[=as.pairlist]{as.pairlist()}}}
\usage{
as_pairlist_signature_linter()
}
\description{
Check signature of \code{\link[=as.pairlist]{as.pairlist()}}
}
