% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_POSIXct_Date_signature_linter.R
\name{as_POSIXct_Date_signature_linter}
\alias{as_POSIXct_Date_signature_linter}
\title{Check signature of \code{\link[=as.POSIXct.Date]{as.POSIXct.Date()}}}
\usage{
as_POSIXct_Date_signature_linter()
}
\description{
Check signature of \code{\link[=as.POSIXct.Date]{as.POSIXct.Date()}}
}
