% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_POSIXlt_default_signature_linter.R
\name{as_POSIXlt_default_signature_linter}
\alias{as_POSIXlt_default_signature_linter}
\title{Check signature of \code{\link[=as.POSIXlt.default]{as.POSIXlt.default()}}}
\usage{
as_POSIXlt_default_signature_linter()
}
\description{
Check signature of \code{\link[=as.POSIXlt.default]{as.POSIXlt.default()}}
}
