% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_qr_signature_linter.R
\name{as_qr_signature_linter}
\alias{as_qr_signature_linter}
\title{Check signature of \code{\link[=as.qr]{as.qr()}}}
\usage{
as_qr_signature_linter()
}
\description{
Check signature of \code{\link[=as.qr]{as.qr()}}
}
