% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_raw_signature_linter.R
\name{as_raw_signature_linter}
\alias{as_raw_signature_linter}
\title{Check signature of \code{\link[=as.raw]{as.raw()}}}
\usage{
as_raw_signature_linter()
}
\description{
Check signature of \code{\link[=as.raw]{as.raw()}}
}
