% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_symbol_signature_linter.R
\name{as_symbol_signature_linter}
\alias{as_symbol_signature_linter}
\title{Check signature of \code{\link[=as.symbol]{as.symbol()}}}
\usage{
as_symbol_signature_linter()
}
\description{
Check signature of \code{\link[=as.symbol]{as.symbol()}}
}
