% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_table_default_signature_linter.R
\name{as_table_default_signature_linter}
\alias{as_table_default_signature_linter}
\title{Check signature of \code{\link[=as.table.default]{as.table.default()}}}
\usage{
as_table_default_signature_linter()
}
\description{
Check signature of \code{\link[=as.table.default]{as.table.default()}}
}
