% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_vector_data_frame_signature_linter.R
\name{as_vector_data_frame_signature_linter}
\alias{as_vector_data_frame_signature_linter}
\title{Check signature of \code{\link[=as.vector.data.frame]{as.vector.data.frame()}}}
\usage{
as_vector_data_frame_signature_linter()
}
\description{
Check signature of \code{\link[=as.vector.data.frame]{as.vector.data.frame()}}
}
