% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asin_signature_linter.R
\name{asin_signature_linter}
\alias{asin_signature_linter}
\title{Check signature of \code{\link[=asin]{asin()}}}
\usage{
asin_signature_linter()
}
\description{
Check signature of \code{\link[=asin]{asin()}}
}
