% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asinh_signature_linter.R
\name{asinh_signature_linter}
\alias{asinh_signature_linter}
\title{Check signature of \code{\link[=asinh]{asinh()}}}
\usage{
asinh_signature_linter()
}
\description{
Check signature of \code{\link[=asinh]{asinh()}}
}
