% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asNamespace_signature_linter.R
\name{asNamespace_signature_linter}
\alias{asNamespace_signature_linter}
\title{Check signature of \code{\link[=asNamespace]{asNamespace()}}}
\usage{
asNamespace_signature_linter()
}
\description{
Check signature of \code{\link[=asNamespace]{asNamespace()}}
}
