% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asplit_signature_linter.R
\name{asplit_signature_linter}
\alias{asplit_signature_linter}
\title{Check signature of \code{\link[=asplit]{asplit()}}}
\usage{
asplit_signature_linter()
}
\description{
Check signature of \code{\link[=asplit]{asplit()}}
}
