% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_signature_linter.R
\name{assign_signature_linter}
\alias{assign_signature_linter}
\title{Check signature of \code{\link[=assign]{assign()}}}
\usage{
assign_signature_linter()
}
\description{
Check signature of \code{\link[=assign]{assign()}}
}
