% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atan_signature_linter.R
\name{atan_signature_linter}
\alias{atan_signature_linter}
\title{Check signature of \code{\link[=atan]{atan()}}}
\usage{
atan_signature_linter()
}
\description{
Check signature of \code{\link[=atan]{atan()}}
}
