% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atanh_signature_linter.R
\name{atanh_signature_linter}
\alias{atanh_signature_linter}
\title{Check signature of \code{\link[=atanh]{atanh()}}}
\usage{
atanh_signature_linter()
}
\description{
Check signature of \code{\link[=atanh]{atanh()}}
}
