% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachNamespace_signature_linter.R
\name{attachNamespace_signature_linter}
\alias{attachNamespace_signature_linter}
\title{Check signature of \code{\link[=attachNamespace]{attachNamespace()}}}
\usage{
attachNamespace_signature_linter()
}
\description{
Check signature of \code{\link[=attachNamespace]{attachNamespace()}}
}
