% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr_signature_linter.R
\name{attr_signature_linter}
\alias{attr_signature_linter}
\title{Check signature of \code{\link[=attr]{attr()}}}
\usage{
attr_signature_linter()
}
\description{
Check signature of \code{\link[=attr]{attr()}}
}
