% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes_signature_linter.R
\name{attributes_signature_linter}
\alias{attributes_signature_linter}
\title{Check signature of \code{\link[=attributes]{attributes()}}}
\usage{
attributes_signature_linter()
}
\description{
Check signature of \code{\link[=attributes]{attributes()}}
}
