% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoload_signature_linter.R
\name{autoload_signature_linter}
\alias{autoload_signature_linter}
\title{Check signature of \code{\link[=autoload]{autoload()}}}
\usage{
autoload_signature_linter()
}
\description{
Check signature of \code{\link[=autoload]{autoload()}}
}
