% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoloader_signature_linter.R
\name{autoloader_signature_linter}
\alias{autoloader_signature_linter}
\title{Check signature of \code{\link[=autoloader]{autoloader()}}}
\usage{
autoloader_signature_linter()
}
\description{
Check signature of \code{\link[=autoloader]{autoloader()}}
}
