% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basename_signature_linter.R
\name{basename_signature_linter}
\alias{basename_signature_linter}
\title{Check signature of \code{\link[=basename]{basename()}}}
\usage{
basename_signature_linter()
}
\description{
Check signature of \code{\link[=basename]{basename()}}
}
