% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_signature_linter.R
\name{beta_signature_linter}
\alias{beta_signature_linter}
\title{Check signature of \code{\link[=beta]{beta()}}}
\usage{
beta_signature_linter()
}
\description{
Check signature of \code{\link[=beta]{beta()}}
}
