% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindingIsActive_signature_linter.R
\name{bindingIsActive_signature_linter}
\alias{bindingIsActive_signature_linter}
\title{Check signature of \code{\link[=bindingIsActive]{bindingIsActive()}}}
\usage{
bindingIsActive_signature_linter()
}
\description{
Check signature of \code{\link[=bindingIsActive]{bindingIsActive()}}
}
