% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindingIsLocked_signature_linter.R
\name{bindingIsLocked_signature_linter}
\alias{bindingIsLocked_signature_linter}
\title{Check signature of \code{\link[=bindingIsLocked]{bindingIsLocked()}}}
\usage{
bindingIsLocked_signature_linter()
}
\description{
Check signature of \code{\link[=bindingIsLocked]{bindingIsLocked()}}
}
