% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindtextdomain_signature_linter.R
\name{bindtextdomain_signature_linter}
\alias{bindtextdomain_signature_linter}
\title{Check signature of \code{\link[=bindtextdomain]{bindtextdomain()}}}
\usage{
bindtextdomain_signature_linter()
}
\description{
Check signature of \code{\link[=bindtextdomain]{bindtextdomain()}}
}
