% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitwAnd_signature_linter.R
\name{bitwAnd_signature_linter}
\alias{bitwAnd_signature_linter}
\title{Check signature of \code{\link[=bitwAnd]{bitwAnd()}}}
\usage{
bitwAnd_signature_linter()
}
\description{
Check signature of \code{\link[=bitwAnd]{bitwAnd()}}
}
