% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitwNot_signature_linter.R
\name{bitwNot_signature_linter}
\alias{bitwNot_signature_linter}
\title{Check signature of \code{\link[=bitwNot]{bitwNot()}}}
\usage{
bitwNot_signature_linter()
}
\description{
Check signature of \code{\link[=bitwNot]{bitwNot()}}
}
