% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitwOr_signature_linter.R
\name{bitwOr_signature_linter}
\alias{bitwOr_signature_linter}
\title{Check signature of \code{\link[=bitwOr]{bitwOr()}}}
\usage{
bitwOr_signature_linter()
}
\description{
Check signature of \code{\link[=bitwOr]{bitwOr()}}
}
