% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitwShiftL_signature_linter.R
\name{bitwShiftL_signature_linter}
\alias{bitwShiftL_signature_linter}
\title{Check signature of \code{\link[=bitwShiftL]{bitwShiftL()}}}
\usage{
bitwShiftL_signature_linter()
}
\description{
Check signature of \code{\link[=bitwShiftL]{bitwShiftL()}}
}
