% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitwShiftR_signature_linter.R
\name{bitwShiftR_signature_linter}
\alias{bitwShiftR_signature_linter}
\title{Check signature of \code{\link[=bitwShiftR]{bitwShiftR()}}}
\usage{
bitwShiftR_signature_linter()
}
\description{
Check signature of \code{\link[=bitwShiftR]{bitwShiftR()}}
}
