% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitwXor_signature_linter.R
\name{bitwXor_signature_linter}
\alias{bitwXor_signature_linter}
\title{Check signature of \code{\link[=bitwXor]{bitwXor()}}}
\usage{
bitwXor_signature_linter()
}
\description{
Check signature of \code{\link[=bitwXor]{bitwXor()}}
}
