% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_signature_linter.R
\name{body_signature_linter}
\alias{body_signature_linter}
\title{Check signature of \code{\link[=body]{body()}}}
\usage{
body_signature_linter()
}
\description{
Check signature of \code{\link[=body]{body()}}
}
