% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bquote_signature_linter.R
\name{bquote_signature_linter}
\alias{bquote_signature_linter}
\title{Check signature of \code{\link[=bquote]{bquote()}}}
\usage{
bquote_signature_linter()
}
\description{
Check signature of \code{\link[=bquote]{bquote()}}
}
