% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_signature_linter.R
\name{browser_signature_linter}
\alias{browser_signature_linter}
\title{Check signature of \code{\link[=browser]{browser()}}}
\usage{
browser_signature_linter()
}
\description{
Check signature of \code{\link[=browser]{browser()}}
}
