% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserText_signature_linter.R
\name{browserText_signature_linter}
\alias{browserText_signature_linter}
\title{Check signature of \code{\link[=browserText]{browserText()}}}
\usage{
browserText_signature_linter()
}
\description{
Check signature of \code{\link[=browserText]{browserText()}}
}
