% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtins_signature_linter.R
\name{builtins_signature_linter}
\alias{builtins_signature_linter}
\title{Check signature of \code{\link[=builtins]{builtins()}}}
\usage{
builtins_signature_linter()
}
\description{
Check signature of \code{\link[=builtins]{builtins()}}
}
