% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bzfile_signature_linter.R
\name{bzfile_signature_linter}
\alias{bzfile_signature_linter}
\title{Check signature of \code{\link[=bzfile]{bzfile()}}}
\usage{
bzfile_signature_linter()
}
\description{
Check signature of \code{\link[=bzfile]{bzfile()}}
}
