% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_noquote_signature_linter.R
\name{c_noquote_signature_linter}
\alias{c_noquote_signature_linter}
\title{Check signature of \code{\link[=c.noquote]{c.noquote()}}}
\usage{
c_noquote_signature_linter()
}
\description{
Check signature of \code{\link[=c.noquote]{c.noquote()}}
}
