% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_warnings_signature_linter.R
\name{c_warnings_signature_linter}
\alias{c_warnings_signature_linter}
\title{Check signature of \code{\link[=c.warnings]{c.warnings()}}}
\usage{
c_warnings_signature_linter()
}
\description{
Check signature of \code{\link[=c.warnings]{c.warnings()}}
}
