% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callCC_signature_linter.R
\name{callCC_signature_linter}
\alias{callCC_signature_linter}
\title{Check signature of \code{\link[=callCC]{callCC()}}}
\usage{
callCC_signature_linter()
}
\description{
Check signature of \code{\link[=callCC]{callCC()}}
}
