% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capabilities_signature_linter.R
\name{capabilities_signature_linter}
\alias{capabilities_signature_linter}
\title{Check signature of \code{\link[=capabilities]{capabilities()}}}
\usage{
capabilities_signature_linter()
}
\description{
Check signature of \code{\link[=capabilities]{capabilities()}}
}
