% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casefold_signature_linter.R
\name{casefold_signature_linter}
\alias{casefold_signature_linter}
\title{Check signature of \code{\link[=casefold]{casefold()}}}
\usage{
casefold_signature_linter()
}
\description{
Check signature of \code{\link[=casefold]{casefold()}}
}
