% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceiling_signature_linter.R
\name{ceiling_signature_linter}
\alias{ceiling_signature_linter}
\title{Check signature of \code{\link[=ceiling]{ceiling()}}}
\usage{
ceiling_signature_linter()
}
\description{
Check signature of \code{\link[=ceiling]{ceiling()}}
}
