% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_expand_signature_linter.R
\name{char_expand_signature_linter}
\alias{char_expand_signature_linter}
\title{Check signature of \code{\link[=char.expand]{char.expand()}}}
\usage{
char_expand_signature_linter()
}
\description{
Check signature of \code{\link[=char.expand]{char.expand()}}
}
