% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_signature_linter.R
\name{character_signature_linter}
\alias{character_signature_linter}
\title{Check signature of \code{\link[=character]{character()}}}
\usage{
character_signature_linter()
}
\description{
Check signature of \code{\link[=character]{character()}}
}
