% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charmatch_signature_linter.R
\name{charmatch_signature_linter}
\alias{charmatch_signature_linter}
\title{Check signature of \code{\link[=charmatch]{charmatch()}}}
\usage{
charmatch_signature_linter()
}
\description{
Check signature of \code{\link[=charmatch]{charmatch()}}
}
