% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charToRaw_signature_linter.R
\name{charToRaw_signature_linter}
\alias{charToRaw_signature_linter}
\title{Check signature of \code{\link[=charToRaw]{charToRaw()}}}
\usage{
charToRaw_signature_linter()
}
\description{
Check signature of \code{\link[=charToRaw]{charToRaw()}}
}
