% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkDots_signature_linter.R
\name{chkDots_signature_linter}
\alias{chkDots_signature_linter}
\title{Check signature of \code{\link[=chkDots]{chkDots()}}}
\usage{
chkDots_signature_linter()
}
\description{
Check signature of \code{\link[=chkDots]{chkDots()}}
}
