% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_signature_linter.R
\name{choose_signature_linter}
\alias{choose_signature_linter}
\title{Check signature of \code{\link[=choose]{choose()}}}
\usage{
choose_signature_linter()
}
\description{
Check signature of \code{\link[=choose]{choose()}}
}
