% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseOpsMethod_default_signature_linter.R
\name{chooseOpsMethod_default_signature_linter}
\alias{chooseOpsMethod_default_signature_linter}
\title{Check signature of \code{\link[=chooseOpsMethod.default]{chooseOpsMethod.default()}}}
\usage{
chooseOpsMethod_default_signature_linter()
}
\description{
Check signature of \code{\link[=chooseOpsMethod.default]{chooseOpsMethod.default()}}
}
