% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseOpsMethod_signature_linter.R
\name{chooseOpsMethod_signature_linter}
\alias{chooseOpsMethod_signature_linter}
\title{Check signature of \code{\link[=chooseOpsMethod]{chooseOpsMethod()}}}
\usage{
chooseOpsMethod_signature_linter()
}
\description{
Check signature of \code{\link[=chooseOpsMethod]{chooseOpsMethod()}}
}
