% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_signature_linter.R
\name{class_signature_linter}
\alias{class_signature_linter}
\title{Check signature of \code{\link[=class]{class()}}}
\usage{
class_signature_linter()
}
\description{
Check signature of \code{\link[=class]{class()}}
}
