% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearPushBack_signature_linter.R
\name{clearPushBack_signature_linter}
\alias{clearPushBack_signature_linter}
\title{Check signature of \code{\link[=clearPushBack]{clearPushBack()}}}
\usage{
clearPushBack_signature_linter()
}
\description{
Check signature of \code{\link[=clearPushBack]{clearPushBack()}}
}
