% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_srcfile_signature_linter.R
\name{close_srcfile_signature_linter}
\alias{close_srcfile_signature_linter}
\title{Check signature of \code{\link[=close.srcfile]{close.srcfile()}}}
\usage{
close_srcfile_signature_linter()
}
\description{
Check signature of \code{\link[=close.srcfile]{close.srcfile()}}
}
