% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_signature_linter.R
\name{col_signature_linter}
\alias{col_signature_linter}
\title{Check signature of \code{\link[=col]{col()}}}
\usage{
col_signature_linter()
}
\description{
Check signature of \code{\link[=col]{col()}}
}
