% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colMeans_signature_linter.R
\name{colMeans_signature_linter}
\alias{colMeans_signature_linter}
\title{Check signature of \code{\link[=colMeans]{colMeans()}}}
\usage{
colMeans_signature_linter()
}
\description{
Check signature of \code{\link[=colMeans]{colMeans()}}
}
