% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_signature_linter.R
\name{colnames_signature_linter}
\alias{colnames_signature_linter}
\title{Check signature of \code{\link[=colnames]{colnames()}}}
\usage{
colnames_signature_linter()
}
\description{
Check signature of \code{\link[=colnames]{colnames()}}
}
