% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colSums_signature_linter.R
\name{colSums_signature_linter}
\alias{colSums_signature_linter}
\title{Check signature of \code{\link[=colSums]{colSums()}}}
\usage{
colSums_signature_linter()
}
\description{
Check signature of \code{\link[=colSums]{colSums()}}
}
