% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandArgs_signature_linter.R
\name{commandArgs_signature_linter}
\alias{commandArgs_signature_linter}
\title{Check signature of \code{\link[=commandArgs]{commandArgs()}}}
\usage{
commandArgs_signature_linter()
}
\description{
Check signature of \code{\link[=commandArgs]{commandArgs()}}
}
