% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_signature_linter.R
\name{comment_signature_linter}
\alias{comment_signature_linter}
\title{Check signature of \code{\link[=comment]{comment()}}}
\usage{
comment_signature_linter()
}
\description{
Check signature of \code{\link[=comment]{comment()}}
}
