% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_signature_linter.R
\name{complex_signature_linter}
\alias{complex_signature_linter}
\title{Check signature of \code{\link[=complex]{complex()}}}
\usage{
complex_signature_linter()
}
\description{
Check signature of \code{\link[=complex]{complex()}}
}
