% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRestarts_signature_linter.R
\name{computeRestarts_signature_linter}
\alias{computeRestarts_signature_linter}
\title{Check signature of \code{\link[=computeRestarts]{computeRestarts()}}}
\usage{
computeRestarts_signature_linter()
}
\description{
Check signature of \code{\link[=computeRestarts]{computeRestarts()}}
}
