% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionCall_condition_signature_linter.R
\name{conditionCall_condition_signature_linter}
\alias{conditionCall_condition_signature_linter}
\title{Check signature of \code{\link[=conditionCall.condition]{conditionCall.condition()}}}
\usage{
conditionCall_condition_signature_linter()
}
\description{
Check signature of \code{\link[=conditionCall.condition]{conditionCall.condition()}}
}
