% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionCall_signature_linter.R
\name{conditionCall_signature_linter}
\alias{conditionCall_signature_linter}
\title{Check signature of \code{\link[=conditionCall]{conditionCall()}}}
\usage{
conditionCall_signature_linter()
}
\description{
Check signature of \code{\link[=conditionCall]{conditionCall()}}
}
