% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionMessage_signature_linter.R
\name{conditionMessage_signature_linter}
\alias{conditionMessage_signature_linter}
\title{Check signature of \code{\link[=conditionMessage]{conditionMessage()}}}
\usage{
conditionMessage_signature_linter()
}
\description{
Check signature of \code{\link[=conditionMessage]{conditionMessage()}}
}
