% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflictRules_signature_linter.R
\name{conflictRules_signature_linter}
\alias{conflictRules_signature_linter}
\title{Check signature of \code{\link[=conflictRules]{conflictRules()}}}
\usage{
conflictRules_signature_linter()
}
\description{
Check signature of \code{\link[=conflictRules]{conflictRules()}}
}
