% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts_signature_linter.R
\name{conflicts_signature_linter}
\alias{conflicts_signature_linter}
\title{Check signature of \code{\link[=conflicts]{conflicts()}}}
\usage{
conflicts_signature_linter()
}
\description{
Check signature of \code{\link[=conflicts]{conflicts()}}
}
