% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conj_signature_linter.R
\name{Conj_signature_linter}
\alias{Conj_signature_linter}
\title{Check signature of \code{\link[=Conj]{Conj()}}}
\usage{
Conj_signature_linter()
}
\description{
Check signature of \code{\link[=Conj]{Conj()}}
}
