% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos_signature_linter.R
\name{cos_signature_linter}
\alias{cos_signature_linter}
\title{Check signature of \code{\link[=cos]{cos()}}}
\usage{
cos_signature_linter()
}
\description{
Check signature of \code{\link[=cos]{cos()}}
}
