% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosh_signature_linter.R
\name{cosh_signature_linter}
\alias{cosh_signature_linter}
\title{Check signature of \code{\link[=cosh]{cosh()}}}
\usage{
cosh_signature_linter()
}
\description{
Check signature of \code{\link[=cosh]{cosh()}}
}
