% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cospi_signature_linter.R
\name{cospi_signature_linter}
\alias{cospi_signature_linter}
\title{Check signature of \code{\link[=cospi]{cospi()}}}
\usage{
cospi_signature_linter()
}
\description{
Check signature of \code{\link[=cospi]{cospi()}}
}
