% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cummax_signature_linter.R
\name{cummax_signature_linter}
\alias{cummax_signature_linter}
\title{Check signature of \code{\link[=cummax]{cummax()}}}
\usage{
cummax_signature_linter()
}
\description{
Check signature of \code{\link[=cummax]{cummax()}}
}
