% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cummin_signature_linter.R
\name{cummin_signature_linter}
\alias{cummin_signature_linter}
\title{Check signature of \code{\link[=cummin]{cummin()}}}
\usage{
cummin_signature_linter()
}
\description{
Check signature of \code{\link[=cummin]{cummin()}}
}
