% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumprod_signature_linter.R
\name{cumprod_signature_linter}
\alias{cumprod_signature_linter}
\title{Check signature of \code{\link[=cumprod]{cumprod()}}}
\usage{
cumprod_signature_linter()
}
\description{
Check signature of \code{\link[=cumprod]{cumprod()}}
}
