% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsum_signature_linter.R
\name{cumsum_signature_linter}
\alias{cumsum_signature_linter}
\title{Check signature of \code{\link[=cumsum]{cumsum()}}}
\usage{
cumsum_signature_linter()
}
\description{
Check signature of \code{\link[=cumsum]{cumsum()}}
}
