% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curlGetHeaders_signature_linter.R
\name{curlGetHeaders_signature_linter}
\alias{curlGetHeaders_signature_linter}
\title{Check signature of \code{\link[=curlGetHeaders]{curlGetHeaders()}}}
\usage{
curlGetHeaders_signature_linter()
}
\description{
Check signature of \code{\link[=curlGetHeaders]{curlGetHeaders()}}
}
