% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_default_signature_linter.R
\name{cut_default_signature_linter}
\alias{cut_default_signature_linter}
\title{Check signature of \code{\link[=cut.default]{cut.default()}}}
\usage{
cut_default_signature_linter()
}
\description{
Check signature of \code{\link[=cut.default]{cut.default()}}
}
