% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_class_signature_linter.R
\name{data_class_signature_linter}
\alias{data_class_signature_linter}
\title{Check signature of \code{\link[=data.class]{data.class()}}}
\usage{
data_class_signature_linter()
}
\description{
Check signature of \code{\link[=data.class]{data.class()}}
}
