% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_matrix_signature_linter.R
\name{data_matrix_signature_linter}
\alias{data_matrix_signature_linter}
\title{Check signature of \code{\link[=data.matrix]{data.matrix()}}}
\usage{
data_matrix_signature_linter()
}
\description{
Check signature of \code{\link[=data.matrix]{data.matrix()}}
}
