% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_signature_linter.R
\name{debug_signature_linter}
\alias{debug_signature_linter}
\title{Check signature of \code{\link[=debug]{debug()}}}
\usage{
debug_signature_linter()
}
\description{
Check signature of \code{\link[=debug]{debug()}}
}
