% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delayedAssign_signature_linter.R
\name{delayedAssign_signature_linter}
\alias{delayedAssign_signature_linter}
\title{Check signature of \code{\link[=delayedAssign]{delayedAssign()}}}
\usage{
delayedAssign_signature_linter()
}
\description{
Check signature of \code{\link[=delayedAssign]{delayedAssign()}}
}
