% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deparse_signature_linter.R
\name{deparse_signature_linter}
\alias{deparse_signature_linter}
\title{Check signature of \code{\link[=deparse]{deparse()}}}
\usage{
deparse_signature_linter()
}
\description{
Check signature of \code{\link[=deparse]{deparse()}}
}
