% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dget_signature_linter.R
\name{dget_signature_linter}
\alias{dget_signature_linter}
\title{Check signature of \code{\link[=dget]{dget()}}}
\usage{
dget_signature_linter()
}
\description{
Check signature of \code{\link[=dget]{dget()}}
}
