% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_signature_linter.R
\name{diag_signature_linter}
\alias{diag_signature_linter}
\title{Check signature of \code{\link[=diag]{diag()}}}
\usage{
diag_signature_linter()
}
\description{
Check signature of \code{\link[=diag]{diag()}}
}
