% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_default_signature_linter.R
\name{diff_default_signature_linter}
\alias{diff_default_signature_linter}
\title{Check signature of \code{\link[=diff.default]{diff.default()}}}
\usage{
diff_default_signature_linter()
}
\description{
Check signature of \code{\link[=diff.default]{diff.default()}}
}
