% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_difftime_signature_linter.R
\name{diff_difftime_signature_linter}
\alias{diff_difftime_signature_linter}
\title{Check signature of \code{\link[=diff.difftime]{diff.difftime()}}}
\usage{
diff_difftime_signature_linter()
}
\description{
Check signature of \code{\link[=diff.difftime]{diff.difftime()}}
}
