% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difftime_signature_linter.R
\name{difftime_signature_linter}
\alias{difftime_signature_linter}
\title{Check signature of \code{\link[=difftime]{difftime()}}}
\usage{
difftime_signature_linter()
}
\description{
Check signature of \code{\link[=difftime]{difftime()}}
}
