% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digamma_signature_linter.R
\name{digamma_signature_linter}
\alias{digamma_signature_linter}
\title{Check signature of \code{\link[=digamma]{digamma()}}}
\usage{
digamma_signature_linter()
}
\description{
Check signature of \code{\link[=digamma]{digamma()}}
}
