% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_data_frame_signature_linter.R
\name{dim_data_frame_signature_linter}
\alias{dim_data_frame_signature_linter}
\title{Check signature of \code{\link[=dim.data.frame]{dim.data.frame()}}}
\usage{
dim_data_frame_signature_linter()
}
\description{
Check signature of \code{\link[=dim.data.frame]{dim.data.frame()}}
}
