% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_signature_linter.R
\name{dim_signature_linter}
\alias{dim_signature_linter}
\title{Check signature of \code{\link[=dim]{dim()}}}
\usage{
dim_signature_linter()
}
\description{
Check signature of \code{\link[=dim]{dim()}}
}
