% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimnames_data_frame_signature_linter.R
\name{dimnames_data_frame_signature_linter}
\alias{dimnames_data_frame_signature_linter}
\title{Check signature of \code{\link[=dimnames.data.frame]{dimnames.data.frame()}}}
\usage{
dimnames_data_frame_signature_linter()
}
\description{
Check signature of \code{\link[=dimnames.data.frame]{dimnames.data.frame()}}
}
