% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimnames_signature_linter.R
\name{dimnames_signature_linter}
\alias{dimnames_signature_linter}
\title{Check signature of \code{\link[=dimnames]{dimnames()}}}
\usage{
dimnames_signature_linter()
}
\description{
Check signature of \code{\link[=dimnames]{dimnames()}}
}
