% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_create_signature_linter.R
\name{dir_create_signature_linter}
\alias{dir_create_signature_linter}
\title{Check signature of \code{\link[=dir.create]{dir.create()}}}
\usage{
dir_create_signature_linter()
}
\description{
Check signature of \code{\link[=dir.create]{dir.create()}}
}
