% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_exists_signature_linter.R
\name{dir_exists_signature_linter}
\alias{dir_exists_signature_linter}
\title{Check signature of \code{\link[=dir.exists]{dir.exists()}}}
\usage{
dir_exists_signature_linter()
}
\description{
Check signature of \code{\link[=dir.exists]{dir.exists()}}
}
