% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirname_signature_linter.R
\name{dirname_signature_linter}
\alias{dirname_signature_linter}
\title{Check signature of \code{\link[=dirname]{dirname()}}}
\usage{
dirname_signature_linter()
}
\description{
Check signature of \code{\link[=dirname]{dirname()}}
}
