% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_call_signature_linter.R
\name{do_call_signature_linter}
\alias{do_call_signature_linter}
\title{Check signature of \code{\link[=do.call]{do.call()}}}
\usage{
do_call_signature_linter()
}
\description{
Check signature of \code{\link[=do.call]{do.call()}}
}
