% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dontCheck_signature_linter.R
\name{dontCheck_signature_linter}
\alias{dontCheck_signature_linter}
\title{Check signature of \code{\link[=dontCheck]{dontCheck()}}}
\usage{
dontCheck_signature_linter()
}
\description{
Check signature of \code{\link[=dontCheck]{dontCheck()}}
}
