% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_signature_linter.R
\name{double_signature_linter}
\alias{double_signature_linter}
\title{Check signature of \code{\link[=double]{double()}}}
\usage{
double_signature_linter()
}
\description{
Check signature of \code{\link[=double]{double()}}
}
