% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dput_signature_linter.R
\name{dput_signature_linter}
\alias{dput_signature_linter}
\title{Check signature of \code{\link[=dput]{dput()}}}
\usage{
dput_signature_linter()
}
\description{
Check signature of \code{\link[=dput]{dput()}}
}
