% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dQuote_signature_linter.R
\name{dQuote_signature_linter}
\alias{dQuote_signature_linter}
\title{Check signature of \code{\link[=dQuote]{dQuote()}}}
\usage{
dQuote_signature_linter()
}
\description{
Check signature of \code{\link[=dQuote]{dQuote()}}
}
