% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_signature_linter.R
\name{drop_signature_linter}
\alias{drop_signature_linter}
\title{Check signature of \code{\link[=drop]{drop()}}}
\usage{
drop_signature_linter()
}
\description{
Check signature of \code{\link[=drop]{drop()}}
}
