% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplevels_data_frame_signature_linter.R
\name{droplevels_data_frame_signature_linter}
\alias{droplevels_data_frame_signature_linter}
\title{Check signature of \code{\link[=droplevels.data.frame]{droplevels.data.frame()}}}
\usage{
droplevels_data_frame_signature_linter()
}
\description{
Check signature of \code{\link[=droplevels.data.frame]{droplevels.data.frame()}}
}
