% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplevels_factor_signature_linter.R
\name{droplevels_factor_signature_linter}
\alias{droplevels_factor_signature_linter}
\title{Check signature of \code{\link[=droplevels.factor]{droplevels.factor()}}}
\usage{
droplevels_factor_signature_linter()
}
\description{
Check signature of \code{\link[=droplevels.factor]{droplevels.factor()}}
}
