% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump_signature_linter.R
\name{dump_signature_linter}
\alias{dump_signature_linter}
\title{Check signature of \code{\link[=dump]{dump()}}}
\usage{
dump_signature_linter()
}
\description{
Check signature of \code{\link[=dump]{dump()}}
}
