% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated_numeric_version_signature_linter.R
\name{duplicated_numeric_version_signature_linter}
\alias{duplicated_numeric_version_signature_linter}
\title{Check signature of \code{\link[=duplicated.numeric_version]{duplicated.numeric_version()}}}
\usage{
duplicated_numeric_version_signature_linter()
}
\description{
Check signature of \code{\link[=duplicated.numeric_version]{duplicated.numeric_version()}}
}
