% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated_POSIXlt_signature_linter.R
\name{duplicated_POSIXlt_signature_linter}
\alias{duplicated_POSIXlt_signature_linter}
\title{Check signature of \code{\link[=duplicated.POSIXlt]{duplicated.POSIXlt()}}}
\usage{
duplicated_POSIXlt_signature_linter()
}
\description{
Check signature of \code{\link[=duplicated.POSIXlt]{duplicated.POSIXlt()}}
}
