% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynGet_signature_linter.R
\name{dynGet_signature_linter}
\alias{dynGet_signature_linter}
\title{Check signature of \code{\link[=dynGet]{dynGet()}}}
\usage{
dynGet_signature_linter()
}
\description{
Check signature of \code{\link[=dynGet]{dynGet()}}
}
