% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_signature_linter.R
\name{eigen_signature_linter}
\alias{eigen_signature_linter}
\title{Check signature of \code{\link[=eigen]{eigen()}}}
\usage{
eigen_signature_linter()
}
\description{
Check signature of \code{\link[=eigen]{eigen()}}
}
