% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeString_signature_linter.R
\name{encodeString_signature_linter}
\alias{encodeString_signature_linter}
\title{Check signature of \code{\link[=encodeString]{encodeString()}}}
\usage{
encodeString_signature_linter()
}
\description{
Check signature of \code{\link[=encodeString]{encodeString()}}
}
