% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encoding_signature_linter.R
\name{Encoding_signature_linter}
\alias{Encoding_signature_linter}
\title{Check signature of \code{\link[=Encoding]{Encoding()}}}
\usage{
Encoding_signature_linter()
}
\description{
Check signature of \code{\link[=Encoding]{Encoding()}}
}
