% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endsWith_signature_linter.R
\name{endsWith_signature_linter}
\alias{endsWith_signature_linter}
\title{Check signature of \code{\link[=endsWith]{endsWith()}}}
\usage{
endsWith_signature_linter()
}
\description{
Check signature of \code{\link[=endsWith]{endsWith()}}
}
