% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_signature_linter.R
\name{environment_signature_linter}
\alias{environment_signature_linter}
\title{Check signature of \code{\link[=environment]{environment()}}}
\usage{
environment_signature_linter()
}
\description{
Check signature of \code{\link[=environment]{environment()}}
}
