% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environmentIsLocked_signature_linter.R
\name{environmentIsLocked_signature_linter}
\alias{environmentIsLocked_signature_linter}
\title{Check signature of \code{\link[=environmentIsLocked]{environmentIsLocked()}}}
\usage{
environmentIsLocked_signature_linter()
}
\description{
Check signature of \code{\link[=environmentIsLocked]{environmentIsLocked()}}
}
