% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environmentName_signature_linter.R
\name{environmentName_signature_linter}
\alias{environmentName_signature_linter}
\title{Check signature of \code{\link[=environmentName]{environmentName()}}}
\usage{
environmentName_signature_linter()
}
\description{
Check signature of \code{\link[=environmentName]{environmentName()}}
}
