% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_parent_signature_linter.R
\name{eval_parent_signature_linter}
\alias{eval_parent_signature_linter}
\title{Check signature of \code{\link[=eval.parent]{eval.parent()}}}
\usage{
eval_parent_signature_linter()
}
\description{
Check signature of \code{\link[=eval.parent]{eval.parent()}}
}
