% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_signature_linter.R
\name{eval_signature_linter}
\alias{eval_signature_linter}
\title{Check signature of \code{\link[=eval]{eval()}}}
\usage{
eval_signature_linter()
}
\description{
Check signature of \code{\link[=eval]{eval()}}
}
