% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalq_signature_linter.R
\name{evalq_signature_linter}
\alias{evalq_signature_linter}
\title{Check signature of \code{\link[=evalq]{evalq()}}}
\usage{
evalq_signature_linter()
}
\description{
Check signature of \code{\link[=evalq]{evalq()}}
}
