% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exec_signature_linter.R
\name{Exec_signature_linter}
\alias{Exec_signature_linter}
\title{Check signature of \code{\link[=Exec]{Exec()}}}
\usage{
Exec_signature_linter()
}
\description{
Check signature of \code{\link[=Exec]{Exec()}}
}
