% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists_signature_linter.R
\name{exists_signature_linter}
\alias{exists_signature_linter}
\title{Check signature of \code{\link[=exists]{exists()}}}
\usage{
exists_signature_linter()
}
\description{
Check signature of \code{\link[=exists]{exists()}}
}
