% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_signature_linter.R
\name{exp_signature_linter}
\alias{exp_signature_linter}
\title{Check signature of \code{\link[=exp]{exp()}}}
\usage{
exp_signature_linter()
}
\description{
Check signature of \code{\link[=exp]{exp()}}
}
