% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_signature_linter.R
\name{factor_signature_linter}
\alias{factor_signature_linter}
\title{Check signature of \code{\link[=factor]{factor()}}}
\usage{
factor_signature_linter()
}
\description{
Check signature of \code{\link[=factor]{factor()}}
}
