% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_signature_linter.R
\name{factorial_signature_linter}
\alias{factorial_signature_linter}
\title{Check signature of \code{\link[=factorial]{factorial()}}}
\usage{
factorial_signature_linter()
}
\description{
Check signature of \code{\link[=factorial]{factorial()}}
}
