% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fifo_signature_linter.R
\name{fifo_signature_linter}
\alias{fifo_signature_linter}
\title{Check signature of \code{\link[=fifo]{fifo()}}}
\usage{
fifo_signature_linter()
}
\description{
Check signature of \code{\link[=fifo]{fifo()}}
}
