% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_access_signature_linter.R
\name{file_access_signature_linter}
\alias{file_access_signature_linter}
\title{Check signature of \code{\link[=file.access]{file.access()}}}
\usage{
file_access_signature_linter()
}
\description{
Check signature of \code{\link[=file.access]{file.access()}}
}
