% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_append_signature_linter.R
\name{file_append_signature_linter}
\alias{file_append_signature_linter}
\title{Check signature of \code{\link[=file.append]{file.append()}}}
\usage{
file_append_signature_linter()
}
\description{
Check signature of \code{\link[=file.append]{file.append()}}
}
