% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_choose_signature_linter.R
\name{file_choose_signature_linter}
\alias{file_choose_signature_linter}
\title{Check signature of \code{\link[=file.choose]{file.choose()}}}
\usage{
file_choose_signature_linter()
}
\description{
Check signature of \code{\link[=file.choose]{file.choose()}}
}
