% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_copy_signature_linter.R
\name{file_copy_signature_linter}
\alias{file_copy_signature_linter}
\title{Check signature of \code{\link[=file.copy]{file.copy()}}}
\usage{
file_copy_signature_linter()
}
\description{
Check signature of \code{\link[=file.copy]{file.copy()}}
}
