% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_link_signature_linter.R
\name{file_link_signature_linter}
\alias{file_link_signature_linter}
\title{Check signature of \code{\link[=file.link]{file.link()}}}
\usage{
file_link_signature_linter()
}
\description{
Check signature of \code{\link[=file.link]{file.link()}}
}
