% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_signature_linter.R
\name{file_signature_linter}
\alias{file_signature_linter}
\title{Check signature of \code{\link[=file]{file()}}}
\usage{
file_signature_linter()
}
\description{
Check signature of \code{\link[=file]{file()}}
}
