% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_symlink_signature_linter.R
\name{file_symlink_signature_linter}
\alias{file_symlink_signature_linter}
\title{Check signature of \code{\link[=file.symlink]{file.symlink()}}}
\usage{
file_symlink_signature_linter()
}
\description{
Check signature of \code{\link[=file.symlink]{file.symlink()}}
}
