% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_signature_linter.R
\name{Filter_signature_linter}
\alias{Filter_signature_linter}
\title{Check signature of \code{\link[=Filter]{Filter()}}}
\usage{
Filter_signature_linter()
}
\description{
Check signature of \code{\link[=Filter]{Filter()}}
}
