% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_package_signature_linter.R
\name{find_package_signature_linter}
\alias{find_package_signature_linter}
\title{Check signature of \code{\link[=find.package]{find.package()}}}
\usage{
find_package_signature_linter()
}
\description{
Check signature of \code{\link[=find.package]{find.package()}}
}
