% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findInterval_signature_linter.R
\name{findInterval_signature_linter}
\alias{findInterval_signature_linter}
\title{Check signature of \code{\link[=findInterval]{findInterval()}}}
\usage{
findInterval_signature_linter()
}
\description{
Check signature of \code{\link[=findInterval]{findInterval()}}
}
