% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPackageEnv_signature_linter.R
\name{findPackageEnv_signature_linter}
\alias{findPackageEnv_signature_linter}
\title{Check signature of \code{\link[=findPackageEnv]{findPackageEnv()}}}
\usage{
findPackageEnv_signature_linter()
}
\description{
Check signature of \code{\link[=findPackageEnv]{findPackageEnv()}}
}
