% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRestart_signature_linter.R
\name{findRestart_signature_linter}
\alias{findRestart_signature_linter}
\title{Check signature of \code{\link[=findRestart]{findRestart()}}}
\usage{
findRestart_signature_linter()
}
\description{
Check signature of \code{\link[=findRestart]{findRestart()}}
}
