% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floor_signature_linter.R
\name{floor_signature_linter}
\alias{floor_signature_linter}
\title{Check signature of \code{\link[=floor]{floor()}}}
\usage{
floor_signature_linter()
}
\description{
Check signature of \code{\link[=floor]{floor()}}
}
