% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flush_connection_signature_linter.R
\name{flush_connection_signature_linter}
\alias{flush_connection_signature_linter}
\title{Check signature of \code{\link[=flush.connection]{flush.connection()}}}
\usage{
flush_connection_signature_linter()
}
\description{
Check signature of \code{\link[=flush.connection]{flush.connection()}}
}
