% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flush_signature_linter.R
\name{flush_signature_linter}
\alias{flush_signature_linter}
\title{Check signature of \code{\link[=flush]{flush()}}}
\usage{
flush_signature_linter()
}
\description{
Check signature of \code{\link[=flush]{flush()}}
}
