% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_signature_linter.R
\name{force_signature_linter}
\alias{force_signature_linter}
\title{Check signature of \code{\link[=force]{force()}}}
\usage{
force_signature_linter()
}
\description{
Check signature of \code{\link[=force]{force()}}
}
