% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formals_signature_linter.R
\name{formals_signature_linter}
\alias{formals_signature_linter}
\title{Check signature of \code{\link[=formals]{formals()}}}
\usage{
formals_signature_linter()
}
\description{
Check signature of \code{\link[=formals]{formals()}}
}
