% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_info_signature_linter.R
\name{format_info_signature_linter}
\alias{format_info_signature_linter}
\title{Check signature of \code{\link[=format.info]{format.info()}}}
\usage{
format_info_signature_linter()
}
\description{
Check signature of \code{\link[=format.info]{format.info()}}
}
