% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numeric_version_signature_linter.R
\name{format_numeric_version_signature_linter}
\alias{format_numeric_version_signature_linter}
\title{Check signature of \code{\link[=format.numeric_version]{format.numeric_version()}}}
\usage{
format_numeric_version_signature_linter()
}
\description{
Check signature of \code{\link[=format.numeric_version]{format.numeric_version()}}
}
