% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_octmode_signature_linter.R
\name{format_octmode_signature_linter}
\alias{format_octmode_signature_linter}
\title{Check signature of \code{\link[=format.octmode]{format.octmode()}}}
\usage{
format_octmode_signature_linter()
}
\description{
Check signature of \code{\link[=format.octmode]{format.octmode()}}
}
