% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_packageInfo_signature_linter.R
\name{format_packageInfo_signature_linter}
\alias{format_packageInfo_signature_linter}
\title{Check signature of \code{\link[=format.packageInfo]{format.packageInfo()}}}
\usage{
format_packageInfo_signature_linter()
}
\description{
Check signature of \code{\link[=format.packageInfo]{format.packageInfo()}}
}
