% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatC_signature_linter.R
\name{formatC_signature_linter}
\alias{formatC_signature_linter}
\title{Check signature of \code{\link[=formatC]{formatC()}}}
\usage{
formatC_signature_linter()
}
\description{
Check signature of \code{\link[=formatC]{formatC()}}
}
