% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatDL_signature_linter.R
\name{formatDL_signature_linter}
\alias{formatDL_signature_linter}
\title{Check signature of \code{\link[=formatDL]{formatDL()}}}
\usage{
formatDL_signature_linter()
}
\description{
Check signature of \code{\link[=formatDL]{formatDL()}}
}
