% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_signature_linter.R
\name{gamma_signature_linter}
\alias{gamma_signature_linter}
\title{Check signature of \code{\link[=gamma]{gamma()}}}
\usage{
gamma_signature_linter()
}
\description{
Check signature of \code{\link[=gamma]{gamma()}}
}
