% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_time_signature_linter.R
\name{gc_time_signature_linter}
\alias{gc_time_signature_linter}
\title{Check signature of \code{\link[=gc.time]{gc.time()}}}
\usage{
gc_time_signature_linter()
}
\description{
Check signature of \code{\link[=gc.time]{gc.time()}}
}
