% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcinfo_signature_linter.R
\name{gcinfo_signature_linter}
\alias{gcinfo_signature_linter}
\title{Check signature of \code{\link[=gcinfo]{gcinfo()}}}
\usage{
gcinfo_signature_linter()
}
\description{
Check signature of \code{\link[=gcinfo]{gcinfo()}}
}
