% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gctorture_signature_linter.R
\name{gctorture_signature_linter}
\alias{gctorture_signature_linter}
\title{Check signature of \code{\link[=gctorture]{gctorture()}}}
\usage{
gctorture_signature_linter()
}
\description{
Check signature of \code{\link[=gctorture]{gctorture()}}
}
