% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signature_linter.R
\name{get_signature_linter}
\alias{get_signature_linter}
\title{Check signature of \code{\link[=get]{get()}}}
\usage{
get_signature_linter()
}
\description{
Check signature of \code{\link[=get]{get()}}
}
