% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCallingDLL_signature_linter.R
\name{getCallingDLL_signature_linter}
\alias{getCallingDLL_signature_linter}
\title{Check signature of \code{\link[=getCallingDLL]{getCallingDLL()}}}
\usage{
getCallingDLL_signature_linter()
}
\description{
Check signature of \code{\link[=getCallingDLL]{getCallingDLL()}}
}
