% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCallingDLLe_signature_linter.R
\name{getCallingDLLe_signature_linter}
\alias{getCallingDLLe_signature_linter}
\title{Check signature of \code{\link[=getCallingDLLe]{getCallingDLLe()}}}
\usage{
getCallingDLLe_signature_linter()
}
\description{
Check signature of \code{\link[=getCallingDLLe]{getCallingDLLe()}}
}
