% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConnection_signature_linter.R
\name{getConnection_signature_linter}
\alias{getConnection_signature_linter}
\title{Check signature of \code{\link[=getConnection]{getConnection()}}}
\usage{
getConnection_signature_linter()
}
\description{
Check signature of \code{\link[=getConnection]{getConnection()}}
}
