% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/getDLLRegisteredRoutines_DLLInfo_signature_linter.R
\name{getDLLRegisteredRoutines_DLLInfo_signature_linter}
\alias{getDLLRegisteredRoutines_DLLInfo_signature_linter}
\title{Check signature of \code{\link[=getDLLRegisteredRoutines.DLLInfo]{getDLLRegisteredRoutines.DLLInfo()}}}
\usage{
getDLLRegisteredRoutines_DLLInfo_signature_linter()
}
\description{
Check signature of \code{\link[=getDLLRegisteredRoutines.DLLInfo]{getDLLRegisteredRoutines.DLLInfo()}}
}
