% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDLLRegisteredRoutines_signature_linter.R
\name{getDLLRegisteredRoutines_signature_linter}
\alias{getDLLRegisteredRoutines_signature_linter}
\title{Check signature of \code{\link[=getDLLRegisteredRoutines]{getDLLRegisteredRoutines()}}}
\usage{
getDLLRegisteredRoutines_signature_linter()
}
\description{
Check signature of \code{\link[=getDLLRegisteredRoutines]{getDLLRegisteredRoutines()}}
}
