% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getElement_signature_linter.R
\name{getElement_signature_linter}
\alias{getElement_signature_linter}
\title{Check signature of \code{\link[=getElement]{getElement()}}}
\usage{
getElement_signature_linter()
}
\description{
Check signature of \code{\link[=getElement]{getElement()}}
}
