% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExportedValue_signature_linter.R
\name{getExportedValue_signature_linter}
\alias{getExportedValue_signature_linter}
\title{Check signature of \code{\link[=getExportedValue]{getExportedValue()}}}
\usage{
getExportedValue_signature_linter()
}
\description{
Check signature of \code{\link[=getExportedValue]{getExportedValue()}}
}
