% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHook_signature_linter.R
\name{getHook_signature_linter}
\alias{getHook_signature_linter}
\title{Check signature of \code{\link[=getHook]{getHook()}}}
\usage{
getHook_signature_linter()
}
\description{
Check signature of \code{\link[=getHook]{getHook()}}
}
