% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNamespace_signature_linter.R
\name{getNamespace_signature_linter}
\alias{getNamespace_signature_linter}
\title{Check signature of \code{\link[=getNamespace]{getNamespace()}}}
\usage{
getNamespace_signature_linter()
}
\description{
Check signature of \code{\link[=getNamespace]{getNamespace()}}
}
