% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNamespaceExports_signature_linter.R
\name{getNamespaceExports_signature_linter}
\alias{getNamespaceExports_signature_linter}
\title{Check signature of \code{\link[=getNamespaceExports]{getNamespaceExports()}}}
\usage{
getNamespaceExports_signature_linter()
}
\description{
Check signature of \code{\link[=getNamespaceExports]{getNamespaceExports()}}
}
