% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNamespaceImports_signature_linter.R
\name{getNamespaceImports_signature_linter}
\alias{getNamespaceImports_signature_linter}
\title{Check signature of \code{\link[=getNamespaceImports]{getNamespaceImports()}}}
\usage{
getNamespaceImports_signature_linter()
}
\description{
Check signature of \code{\link[=getNamespaceImports]{getNamespaceImports()}}
}
