% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNamespaceInfo_signature_linter.R
\name{getNamespaceInfo_signature_linter}
\alias{getNamespaceInfo_signature_linter}
\title{Check signature of \code{\link[=getNamespaceInfo]{getNamespaceInfo()}}}
\usage{
getNamespaceInfo_signature_linter()
}
\description{
Check signature of \code{\link[=getNamespaceInfo]{getNamespaceInfo()}}
}
