% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNamespaceName_signature_linter.R
\name{getNamespaceName_signature_linter}
\alias{getNamespaceName_signature_linter}
\title{Check signature of \code{\link[=getNamespaceName]{getNamespaceName()}}}
\usage{
getNamespaceName_signature_linter()
}
\description{
Check signature of \code{\link[=getNamespaceName]{getNamespaceName()}}
}
