% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNamespaceUsers_signature_linter.R
\name{getNamespaceUsers_signature_linter}
\alias{getNamespaceUsers_signature_linter}
\title{Check signature of \code{\link[=getNamespaceUsers]{getNamespaceUsers()}}}
\usage{
getNamespaceUsers_signature_linter()
}
\description{
Check signature of \code{\link[=getNamespaceUsers]{getNamespaceUsers()}}
}
