% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNamespaceVersion_signature_linter.R
\name{getNamespaceVersion_signature_linter}
\alias{getNamespaceVersion_signature_linter}
\title{Check signature of \code{\link[=getNamespaceVersion]{getNamespaceVersion()}}}
\usage{
getNamespaceVersion_signature_linter()
}
\description{
Check signature of \code{\link[=getNamespaceVersion]{getNamespaceVersion()}}
}
