% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNativeSymbolInfo_signature_linter.R
\name{getNativeSymbolInfo_signature_linter}
\alias{getNativeSymbolInfo_signature_linter}
\title{Check signature of \code{\link[=getNativeSymbolInfo]{getNativeSymbolInfo()}}}
\usage{
getNativeSymbolInfo_signature_linter()
}
\description{
Check signature of \code{\link[=getNativeSymbolInfo]{getNativeSymbolInfo()}}
}
