% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOption_signature_linter.R
\name{getOption_signature_linter}
\alias{getOption_signature_linter}
\title{Check signature of \code{\link[=getOption]{getOption()}}}
\usage{
getOption_signature_linter()
}
\description{
Check signature of \code{\link[=getOption]{getOption()}}
}
