% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSrcLines_signature_linter.R
\name{getSrcLines_signature_linter}
\alias{getSrcLines_signature_linter}
\title{Check signature of \code{\link[=getSrcLines]{getSrcLines()}}}
\usage{
getSrcLines_signature_linter()
}
\description{
Check signature of \code{\link[=getSrcLines]{getSrcLines()}}
}
