% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregexec_signature_linter.R
\name{gregexec_signature_linter}
\alias{gregexec_signature_linter}
\title{Check signature of \code{\link[=gregexec]{gregexec()}}}
\usage{
gregexec_signature_linter()
}
\description{
Check signature of \code{\link[=gregexec]{gregexec()}}
}
