% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregexpr_signature_linter.R
\name{gregexpr_signature_linter}
\alias{gregexpr_signature_linter}
\title{Check signature of \code{\link[=gregexpr]{gregexpr()}}}
\usage{
gregexpr_signature_linter()
}
\description{
Check signature of \code{\link[=gregexpr]{gregexpr()}}
}
