% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep_signature_linter.R
\name{grep_signature_linter}
\alias{grep_signature_linter}
\title{Check signature of \code{\link[=grep]{grep()}}}
\usage{
grep_signature_linter()
}
\description{
Check signature of \code{\link[=grep]{grep()}}
}
