% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepl_signature_linter.R
\name{grepl_signature_linter}
\alias{grepl_signature_linter}
\title{Check signature of \code{\link[=grepl]{grepl()}}}
\usage{
grepl_signature_linter()
}
\description{
Check signature of \code{\link[=grepl]{grepl()}}
}
