% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepRaw_signature_linter.R
\name{grepRaw_signature_linter}
\alias{grepRaw_signature_linter}
\title{Check signature of \code{\link[=grepRaw]{grepRaw()}}}
\usage{
grepRaw_signature_linter()
}
\description{
Check signature of \code{\link[=grepRaw]{grepRaw()}}
}
