% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepv_signature_linter.R
\name{grepv_signature_linter}
\alias{grepv_signature_linter}
\title{Check signature of \code{\link[=grepv]{grepv()}}}
\usage{
grepv_signature_linter()
}
\description{
Check signature of \code{\link[=grepv]{grepv()}}
}
