% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsub_signature_linter.R
\name{gsub_signature_linter}
\alias{gsub_signature_linter}
\title{Check signature of \code{\link[=gsub]{gsub()}}}
\usage{
gsub_signature_linter()
}
\description{
Check signature of \code{\link[=gsub]{gsub()}}
}
