% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gzcon_signature_linter.R
\name{gzcon_signature_linter}
\alias{gzcon_signature_linter}
\title{Check signature of \code{\link[=gzcon]{gzcon()}}}
\usage{
gzcon_signature_linter()
}
\description{
Check signature of \code{\link[=gzcon]{gzcon()}}
}
