% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gzfile_signature_linter.R
\name{gzfile_signature_linter}
\alias{gzfile_signature_linter}
\title{Check signature of \code{\link[=gzfile]{gzfile()}}}
\usage{
gzfile_signature_linter()
}
\description{
Check signature of \code{\link[=gzfile]{gzfile()}}
}
