% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icuGetCollate_signature_linter.R
\name{icuGetCollate_signature_linter}
\alias{icuGetCollate_signature_linter}
\title{Check signature of \code{\link[=icuGetCollate]{icuGetCollate()}}}
\usage{
icuGetCollate_signature_linter()
}
\description{
Check signature of \code{\link[=icuGetCollate]{icuGetCollate()}}
}
