% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identical_signature_linter.R
\name{identical_signature_linter}
\alias{identical_signature_linter}
\title{Check signature of \code{\link[=identical]{identical()}}}
\usage{
identical_signature_linter()
}
\description{
Check signature of \code{\link[=identical]{identical()}}
}
