% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity_signature_linter.R
\name{identity_signature_linter}
\alias{identity_signature_linter}
\title{Check signature of \code{\link[=identity]{identity()}}}
\usage{
identity_signature_linter()
}
\description{
Check signature of \code{\link[=identity]{identity()}}
}
