% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifelse_signature_linter.R
\name{ifelse_signature_linter}
\alias{ifelse_signature_linter}
\title{Check signature of \code{\link[=ifelse]{ifelse()}}}
\usage{
ifelse_signature_linter()
}
\description{
Check signature of \code{\link[=ifelse]{ifelse()}}
}
