% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importIntoEnv_signature_linter.R
\name{importIntoEnv_signature_linter}
\alias{importIntoEnv_signature_linter}
\title{Check signature of \code{\link[=importIntoEnv]{importIntoEnv()}}}
\usage{
importIntoEnv_signature_linter()
}
\description{
Check signature of \code{\link[=importIntoEnv]{importIntoEnv()}}
}
