% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoRDS_signature_linter.R
\name{infoRDS_signature_linter}
\alias{infoRDS_signature_linter}
\title{Check signature of \code{\link[=infoRDS]{infoRDS()}}}
\usage{
infoRDS_signature_linter()
}
\description{
Check signature of \code{\link[=infoRDS]{infoRDS()}}
}
