% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherits_signature_linter.R
\name{inherits_signature_linter}
\alias{inherits_signature_linter}
\title{Check signature of \code{\link[=inherits]{inherits()}}}
\usage{
inherits_signature_linter()
}
\description{
Check signature of \code{\link[=inherits]{inherits()}}
}
