% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer_signature_linter.R
\name{integer_signature_linter}
\alias{integer_signature_linter}
\title{Check signature of \code{\link[=integer]{integer()}}}
\usage{
integer_signature_linter()
}
\description{
Check signature of \code{\link[=integer]{integer()}}
}
