% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_signature_linter.R
\name{intersect_signature_linter}
\alias{intersect_signature_linter}
\title{Check signature of \code{\link[=intersect]{intersect()}}}
\usage{
intersect_signature_linter()
}
\description{
Check signature of \code{\link[=intersect]{intersect()}}
}
