% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intToBits_signature_linter.R
\name{intToBits_signature_linter}
\alias{intToBits_signature_linter}
\title{Check signature of \code{\link[=intToBits]{intToBits()}}}
\usage{
intToBits_signature_linter()
}
\description{
Check signature of \code{\link[=intToBits]{intToBits()}}
}
