% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_rle_signature_linter.R
\name{inverse_rle_signature_linter}
\alias{inverse_rle_signature_linter}
\title{Check signature of \code{\link[=inverse.rle]{inverse.rle()}}}
\usage{
inverse_rle_signature_linter()
}
\description{
Check signature of \code{\link[=inverse.rle]{inverse.rle()}}
}
