% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invisible_signature_linter.R
\name{invisible_signature_linter}
\alias{invisible_signature_linter}
\title{Check signature of \code{\link[=invisible]{invisible()}}}
\usage{
invisible_signature_linter()
}
\description{
Check signature of \code{\link[=invisible]{invisible()}}
}
