% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invokeRestartInteractively_signature_linter.R
\name{invokeRestartInteractively_signature_linter}
\alias{invokeRestartInteractively_signature_linter}
\title{Check signature of \code{\link[=invokeRestartInteractively]{invokeRestartInteractively()}}}
\usage{
invokeRestartInteractively_signature_linter()
}
\description{
Check signature of \code{\link[=invokeRestartInteractively]{invokeRestartInteractively()}}
}
