% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_array_signature_linter.R
\name{is_array_signature_linter}
\alias{is_array_signature_linter}
\title{Check signature of \code{\link[=is.array]{is.array()}}}
\usage{
is_array_signature_linter()
}
\description{
Check signature of \code{\link[=is.array]{is.array()}}
}
