% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_atomic_signature_linter.R
\name{is_atomic_signature_linter}
\alias{is_atomic_signature_linter}
\title{Check signature of \code{\link[=is.atomic]{is.atomic()}}}
\usage{
is_atomic_signature_linter()
}
\description{
Check signature of \code{\link[=is.atomic]{is.atomic()}}
}
