% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_call_signature_linter.R
\name{is_call_signature_linter}
\alias{is_call_signature_linter}
\title{Check signature of \code{\link[=is.call]{is.call()}}}
\usage{
is_call_signature_linter()
}
\description{
Check signature of \code{\link[=is.call]{is.call()}}
}
