% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_character_signature_linter.R
\name{is_character_signature_linter}
\alias{is_character_signature_linter}
\title{Check signature of \code{\link[=is.character]{is.character()}}}
\usage{
is_character_signature_linter()
}
\description{
Check signature of \code{\link[=is.character]{is.character()}}
}
