% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_complex_signature_linter.R
\name{is_complex_signature_linter}
\alias{is_complex_signature_linter}
\title{Check signature of \code{\link[=is.complex]{is.complex()}}}
\usage{
is_complex_signature_linter()
}
\description{
Check signature of \code{\link[=is.complex]{is.complex()}}
}
