% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_data_frame_signature_linter.R
\name{is_data_frame_signature_linter}
\alias{is_data_frame_signature_linter}
\title{Check signature of \code{\link[=is.data.frame]{is.data.frame()}}}
\usage{
is_data_frame_signature_linter()
}
\description{
Check signature of \code{\link[=is.data.frame]{is.data.frame()}}
}
