% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_double_signature_linter.R
\name{is_double_signature_linter}
\alias{is_double_signature_linter}
\title{Check signature of \code{\link[=is.double]{is.double()}}}
\usage{
is_double_signature_linter()
}
\description{
Check signature of \code{\link[=is.double]{is.double()}}
}
