% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_element_signature_linter.R
\name{is_element_signature_linter}
\alias{is_element_signature_linter}
\title{Check signature of \code{\link[=is.element]{is.element()}}}
\usage{
is_element_signature_linter()
}
\description{
Check signature of \code{\link[=is.element]{is.element()}}
}
