% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_environment_signature_linter.R
\name{is_environment_signature_linter}
\alias{is_environment_signature_linter}
\title{Check signature of \code{\link[=is.environment]{is.environment()}}}
\usage{
is_environment_signature_linter()
}
\description{
Check signature of \code{\link[=is.environment]{is.environment()}}
}
