% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_expression_signature_linter.R
\name{is_expression_signature_linter}
\alias{is_expression_signature_linter}
\title{Check signature of \code{\link[=is.expression]{is.expression()}}}
\usage{
is_expression_signature_linter()
}
\description{
Check signature of \code{\link[=is.expression]{is.expression()}}
}
