% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_factor_signature_linter.R
\name{is_factor_signature_linter}
\alias{is_factor_signature_linter}
\title{Check signature of \code{\link[=is.factor]{is.factor()}}}
\usage{
is_factor_signature_linter()
}
\description{
Check signature of \code{\link[=is.factor]{is.factor()}}
}
