% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_finite_POSIXlt_signature_linter.R
\name{is_finite_POSIXlt_signature_linter}
\alias{is_finite_POSIXlt_signature_linter}
\title{Check signature of \code{\link[=is.finite.POSIXlt]{is.finite.POSIXlt()}}}
\usage{
is_finite_POSIXlt_signature_linter()
}
\description{
Check signature of \code{\link[=is.finite.POSIXlt]{is.finite.POSIXlt()}}
}
