% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_function_signature_linter.R
\name{is_function_signature_linter}
\alias{is_function_signature_linter}
\title{Check signature of \code{\link[=is.function]{is.function()}}}
\usage{
is_function_signature_linter()
}
\description{
Check signature of \code{\link[=is.function]{is.function()}}
}
