% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_infinite_POSIXlt_signature_linter.R
\name{is_infinite_POSIXlt_signature_linter}
\alias{is_infinite_POSIXlt_signature_linter}
\title{Check signature of \code{\link[=is.infinite.POSIXlt]{is.infinite.POSIXlt()}}}
\usage{
is_infinite_POSIXlt_signature_linter()
}
\description{
Check signature of \code{\link[=is.infinite.POSIXlt]{is.infinite.POSIXlt()}}
}
