% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_integer_signature_linter.R
\name{is_integer_signature_linter}
\alias{is_integer_signature_linter}
\title{Check signature of \code{\link[=is.integer]{is.integer()}}}
\usage{
is_integer_signature_linter()
}
\description{
Check signature of \code{\link[=is.integer]{is.integer()}}
}
