% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_language_signature_linter.R
\name{is_language_signature_linter}
\alias{is_language_signature_linter}
\title{Check signature of \code{\link[=is.language]{is.language()}}}
\usage{
is_language_signature_linter()
}
\description{
Check signature of \code{\link[=is.language]{is.language()}}
}
