% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_list_signature_linter.R
\name{is_list_signature_linter}
\alias{is_list_signature_linter}
\title{Check signature of \code{\link[=is.list]{is.list()}}}
\usage{
is_list_signature_linter()
}
\description{
Check signature of \code{\link[=is.list]{is.list()}}
}
