% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_loaded_signature_linter.R
\name{is_loaded_signature_linter}
\alias{is_loaded_signature_linter}
\title{Check signature of \code{\link[=is.loaded]{is.loaded()}}}
\usage{
is_loaded_signature_linter()
}
\description{
Check signature of \code{\link[=is.loaded]{is.loaded()}}
}
