% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_logical_signature_linter.R
\name{is_logical_signature_linter}
\alias{is_logical_signature_linter}
\title{Check signature of \code{\link[=is.logical]{is.logical()}}}
\usage{
is_logical_signature_linter()
}
\description{
Check signature of \code{\link[=is.logical]{is.logical()}}
}
