% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_matrix_signature_linter.R
\name{is_matrix_signature_linter}
\alias{is_matrix_signature_linter}
\title{Check signature of \code{\link[=is.matrix]{is.matrix()}}}
\usage{
is_matrix_signature_linter()
}
\description{
Check signature of \code{\link[=is.matrix]{is.matrix()}}
}
