% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_na_data_frame_signature_linter.R
\name{is_na_data_frame_signature_linter}
\alias{is_na_data_frame_signature_linter}
\title{Check signature of \code{\link[=is.na.data.frame]{is.na.data.frame()}}}
\usage{
is_na_data_frame_signature_linter()
}
\description{
Check signature of \code{\link[=is.na.data.frame]{is.na.data.frame()}}
}
