% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_name_signature_linter.R
\name{is_name_signature_linter}
\alias{is_name_signature_linter}
\title{Check signature of \code{\link[=is.name]{is.name()}}}
\usage{
is_name_signature_linter()
}
\description{
Check signature of \code{\link[=is.name]{is.name()}}
}
