% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_nan_signature_linter.R
\name{is_nan_signature_linter}
\alias{is_nan_signature_linter}
\title{Check signature of \code{\link[=is.nan]{is.nan()}}}
\usage{
is_nan_signature_linter()
}
\description{
Check signature of \code{\link[=is.nan]{is.nan()}}
}
