% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_null_signature_linter.R
\name{is_null_signature_linter}
\alias{is_null_signature_linter}
\title{Check signature of \code{\link[=is.null]{is.null()}}}
\usage{
is_null_signature_linter()
}
\description{
Check signature of \code{\link[=is.null]{is.null()}}
}
