% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_numeric_Date_signature_linter.R
\name{is_numeric_Date_signature_linter}
\alias{is_numeric_Date_signature_linter}
\title{Check signature of \code{\link[=is.numeric.Date]{is.numeric.Date()}}}
\usage{
is_numeric_Date_signature_linter()
}
\description{
Check signature of \code{\link[=is.numeric.Date]{is.numeric.Date()}}
}
