% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_numeric_difftime_signature_linter.R
\name{is_numeric_difftime_signature_linter}
\alias{is_numeric_difftime_signature_linter}
\title{Check signature of \code{\link[=is.numeric.difftime]{is.numeric.difftime()}}}
\usage{
is_numeric_difftime_signature_linter()
}
\description{
Check signature of \code{\link[=is.numeric.difftime]{is.numeric.difftime()}}
}
