% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_object_signature_linter.R
\name{is_object_signature_linter}
\alias{is_object_signature_linter}
\title{Check signature of \code{\link[=is.object]{is.object()}}}
\usage{
is_object_signature_linter()
}
\description{
Check signature of \code{\link[=is.object]{is.object()}}
}
