% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ordered_signature_linter.R
\name{is_ordered_signature_linter}
\alias{is_ordered_signature_linter}
\title{Check signature of \code{\link[=is.ordered]{is.ordered()}}}
\usage{
is_ordered_signature_linter()
}
\description{
Check signature of \code{\link[=is.ordered]{is.ordered()}}
}
