% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_package_version_signature_linter.R
\name{is_package_version_signature_linter}
\alias{is_package_version_signature_linter}
\title{Check signature of \code{\link[=is.package_version]{is.package_version()}}}
\usage{
is_package_version_signature_linter()
}
\description{
Check signature of \code{\link[=is.package_version]{is.package_version()}}
}
