% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pairlist_signature_linter.R
\name{is_pairlist_signature_linter}
\alias{is_pairlist_signature_linter}
\title{Check signature of \code{\link[=is.pairlist]{is.pairlist()}}}
\usage{
is_pairlist_signature_linter()
}
\description{
Check signature of \code{\link[=is.pairlist]{is.pairlist()}}
}
