% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_primitive_signature_linter.R
\name{is_primitive_signature_linter}
\alias{is_primitive_signature_linter}
\title{Check signature of \code{\link[=is.primitive]{is.primitive()}}}
\usage{
is_primitive_signature_linter()
}
\description{
Check signature of \code{\link[=is.primitive]{is.primitive()}}
}
