% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_qr_signature_linter.R
\name{is_qr_signature_linter}
\alias{is_qr_signature_linter}
\title{Check signature of \code{\link[=is.qr]{is.qr()}}}
\usage{
is_qr_signature_linter()
}
\description{
Check signature of \code{\link[=is.qr]{is.qr()}}
}
