% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_raw_signature_linter.R
\name{is_raw_signature_linter}
\alias{is_raw_signature_linter}
\title{Check signature of \code{\link[=is.raw]{is.raw()}}}
\usage{
is_raw_signature_linter()
}
\description{
Check signature of \code{\link[=is.raw]{is.raw()}}
}
