% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_recursive_signature_linter.R
\name{is_recursive_signature_linter}
\alias{is_recursive_signature_linter}
\title{Check signature of \code{\link[=is.recursive]{is.recursive()}}}
\usage{
is_recursive_signature_linter()
}
\description{
Check signature of \code{\link[=is.recursive]{is.recursive()}}
}
