% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_single_signature_linter.R
\name{is_single_signature_linter}
\alias{is_single_signature_linter}
\title{Check signature of \code{\link[=is.single]{is.single()}}}
\usage{
is_single_signature_linter()
}
\description{
Check signature of \code{\link[=is.single]{is.single()}}
}
