% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_symbol_signature_linter.R
\name{is_symbol_signature_linter}
\alias{is_symbol_signature_linter}
\title{Check signature of \code{\link[=is.symbol]{is.symbol()}}}
\usage{
is_symbol_signature_linter()
}
\description{
Check signature of \code{\link[=is.symbol]{is.symbol()}}
}
