% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_table_signature_linter.R
\name{is_table_signature_linter}
\alias{is_table_signature_linter}
\title{Check signature of \code{\link[=is.table]{is.table()}}}
\usage{
is_table_signature_linter()
}
\description{
Check signature of \code{\link[=is.table]{is.table()}}
}
