% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unsorted_signature_linter.R
\name{is_unsorted_signature_linter}
\alias{is_unsorted_signature_linter}
\title{Check signature of \code{\link[=is.unsorted]{is.unsorted()}}}
\usage{
is_unsorted_signature_linter()
}
\description{
Check signature of \code{\link[=is.unsorted]{is.unsorted()}}
}
