% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_vector_signature_linter.R
\name{is_vector_signature_linter}
\alias{is_vector_signature_linter}
\title{Check signature of \code{\link[=is.vector]{is.vector()}}}
\usage{
is_vector_signature_linter()
}
\description{
Check signature of \code{\link[=is.vector]{is.vector()}}
}
