% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isa_signature_linter.R
\name{isa_signature_linter}
\alias{isa_signature_linter}
\title{Check signature of \code{\link[=isa]{isa()}}}
\usage{
isa_signature_linter()
}
\description{
Check signature of \code{\link[=isa]{isa()}}
}
