% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatty_signature_linter.R
\name{isatty_signature_linter}
\alias{isatty_signature_linter}
\title{Check signature of \code{\link[=isatty]{isatty()}}}
\usage{
isatty_signature_linter()
}
\description{
Check signature of \code{\link[=isatty]{isatty()}}
}
