% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isBaseNamespace_signature_linter.R
\name{isBaseNamespace_signature_linter}
\alias{isBaseNamespace_signature_linter}
\title{Check signature of \code{\link[=isBaseNamespace]{isBaseNamespace()}}}
\usage{
isBaseNamespace_signature_linter()
}
\description{
Check signature of \code{\link[=isBaseNamespace]{isBaseNamespace()}}
}
