% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isdebugged_signature_linter.R
\name{isdebugged_signature_linter}
\alias{isdebugged_signature_linter}
\title{Check signature of \code{\link[=isdebugged]{isdebugged()}}}
\usage{
isdebugged_signature_linter()
}
\description{
Check signature of \code{\link[=isdebugged]{isdebugged()}}
}
