% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFALSE_signature_linter.R
\name{isFALSE_signature_linter}
\alias{isFALSE_signature_linter}
\title{Check signature of \code{\link[=isFALSE]{isFALSE()}}}
\usage{
isFALSE_signature_linter()
}
\description{
Check signature of \code{\link[=isFALSE]{isFALSE()}}
}
