% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isNamespace_signature_linter.R
\name{isNamespace_signature_linter}
\alias{isNamespace_signature_linter}
\title{Check signature of \code{\link[=isNamespace]{isNamespace()}}}
\usage{
isNamespace_signature_linter()
}
\description{
Check signature of \code{\link[=isNamespace]{isNamespace()}}
}
