% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isNamespaceLoaded_signature_linter.R
\name{isNamespaceLoaded_signature_linter}
\alias{isNamespaceLoaded_signature_linter}
\title{Check signature of \code{\link[=isNamespaceLoaded]{isNamespaceLoaded()}}}
\usage{
isNamespaceLoaded_signature_linter()
}
\description{
Check signature of \code{\link[=isNamespaceLoaded]{isNamespaceLoaded()}}
}
