% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOdatetime_signature_linter.R
\name{ISOdatetime_signature_linter}
\alias{ISOdatetime_signature_linter}
\title{Check signature of \code{\link[=ISOdatetime]{ISOdatetime()}}}
\usage{
ISOdatetime_signature_linter()
}
\description{
Check signature of \code{\link[=ISOdatetime]{ISOdatetime()}}
}
