% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isOpen_signature_linter.R
\name{isOpen_signature_linter}
\alias{isOpen_signature_linter}
\title{Check signature of \code{\link[=isOpen]{isOpen()}}}
\usage{
isOpen_signature_linter()
}
\description{
Check signature of \code{\link[=isOpen]{isOpen()}}
}
