% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isRestart_signature_linter.R
\name{isRestart_signature_linter}
\alias{isRestart_signature_linter}
\title{Check signature of \code{\link[=isRestart]{isRestart()}}}
\usage{
isRestart_signature_linter()
}
\description{
Check signature of \code{\link[=isRestart]{isRestart()}}
}
