% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSeekable_signature_linter.R
\name{isSeekable_signature_linter}
\alias{isSeekable_signature_linter}
\title{Check signature of \code{\link[=isSeekable]{isSeekable()}}}
\usage{
isSeekable_signature_linter()
}
\description{
Check signature of \code{\link[=isSeekable]{isSeekable()}}
}
