% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isTRUE_signature_linter.R
\name{isTRUE_signature_linter}
\alias{isTRUE_signature_linter}
\title{Check signature of \code{\link[=isTRUE]{isTRUE()}}}
\usage{
isTRUE_signature_linter()
}
\description{
Check signature of \code{\link[=isTRUE]{isTRUE()}}
}
