% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter_signature_linter.R
\name{jitter_signature_linter}
\alias{jitter_signature_linter}
\title{Check signature of \code{\link[=jitter]{jitter()}}}
\usage{
jitter_signature_linter()
}
\description{
Check signature of \code{\link[=jitter]{jitter()}}
}
