% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian_Date_signature_linter.R
\name{julian_Date_signature_linter}
\alias{julian_Date_signature_linter}
\title{Check signature of \code{\link[=julian.Date]{julian.Date()}}}
\usage{
julian_Date_signature_linter()
}
\description{
Check signature of \code{\link[=julian.Date]{julian.Date()}}
}
