% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian_POSIXt_signature_linter.R
\name{julian_POSIXt_signature_linter}
\alias{julian_POSIXt_signature_linter}
\title{Check signature of \code{\link[=julian.POSIXt]{julian.POSIXt()}}}
\usage{
julian_POSIXt_signature_linter()
}
\description{
Check signature of \code{\link[=julian.POSIXt]{julian.POSIXt()}}
}
