% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronecker_signature_linter.R
\name{kronecker_signature_linter}
\alias{kronecker_signature_linter}
\title{Check signature of \code{\link[=kronecker]{kronecker()}}}
\usage{
kronecker_signature_linter()
}
\description{
Check signature of \code{\link[=kronecker]{kronecker()}}
}
