% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/La_svd_signature_linter.R
\name{La_svd_signature_linter}
\alias{La_svd_signature_linter}
\title{Check signature of \code{\link[=La.svd]{La.svd()}}}
\usage{
La_svd_signature_linter()
}
\description{
Check signature of \code{\link[=La.svd]{La.svd()}}
}
