% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels_default_signature_linter.R
\name{labels_default_signature_linter}
\alias{labels_default_signature_linter}
\title{Check signature of \code{\link[=labels.default]{labels.default()}}}
\usage{
labels_default_signature_linter()
}
\description{
Check signature of \code{\link[=labels.default]{labels.default()}}
}
