% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbeta_signature_linter.R
\name{lbeta_signature_linter}
\alias{lbeta_signature_linter}
\title{Check signature of \code{\link[=lbeta]{lbeta()}}}
\usage{
lbeta_signature_linter()
}
\description{
Check signature of \code{\link[=lbeta]{lbeta()}}
}
