% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lchoose_signature_linter.R
\name{lchoose_signature_linter}
\alias{lchoose_signature_linter}
\title{Check signature of \code{\link[=lchoose]{lchoose()}}}
\usage{
lchoose_signature_linter()
}
\description{
Check signature of \code{\link[=lchoose]{lchoose()}}
}
