% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_signature_linter.R
\name{length_signature_linter}
\alias{length_signature_linter}
\title{Check signature of \code{\link[=length]{length()}}}
\usage{
length_signature_linter()
}
\description{
Check signature of \code{\link[=length]{length()}}
}
