% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengths_signature_linter.R
\name{lengths_signature_linter}
\alias{lengths_signature_linter}
\title{Check signature of \code{\link[=lengths]{lengths()}}}
\usage{
lengths_signature_linter()
}
\description{
Check signature of \code{\link[=lengths]{lengths()}}
}
