% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels_default_signature_linter.R
\name{levels_default_signature_linter}
\alias{levels_default_signature_linter}
\title{Check signature of \code{\link[=levels.default]{levels.default()}}}
\usage{
levels_default_signature_linter()
}
\description{
Check signature of \code{\link[=levels.default]{levels.default()}}
}
