% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels_signature_linter.R
\name{levels_signature_linter}
\alias{levels_signature_linter}
\title{Check signature of \code{\link[=levels]{levels()}}}
\usage{
levels_signature_linter()
}
\description{
Check signature of \code{\link[=levels]{levels()}}
}
