% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfactorial_signature_linter.R
\name{lfactorial_signature_linter}
\alias{lfactorial_signature_linter}
\title{Check signature of \code{\link[=lfactorial]{lfactorial()}}}
\usage{
lfactorial_signature_linter()
}
\description{
Check signature of \code{\link[=lfactorial]{lfactorial()}}
}
