% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgamma_signature_linter.R
\name{lgamma_signature_linter}
\alias{lgamma_signature_linter}
\title{Check signature of \code{\link[=lgamma]{lgamma()}}}
\usage{
lgamma_signature_linter()
}
\description{
Check signature of \code{\link[=lgamma]{lgamma()}}
}
