% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_dirs_signature_linter.R
\name{list_dirs_signature_linter}
\alias{list_dirs_signature_linter}
\title{Check signature of \code{\link[=list.dirs]{list.dirs()}}}
\usage{
list_dirs_signature_linter()
}
\description{
Check signature of \code{\link[=list.dirs]{list.dirs()}}
}
