% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files_signature_linter.R
\name{list_files_signature_linter}
\alias{list_files_signature_linter}
\title{Check signature of \code{\link[=list.files]{list.files()}}}
\usage{
list_files_signature_linter()
}
\description{
Check signature of \code{\link[=list.files]{list.files()}}
}
