% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_signature_linter.R
\name{load_signature_linter}
\alias{load_signature_linter}
\title{Check signature of \code{\link[=load]{load()}}}
\usage{
load_signature_linter()
}
\description{
Check signature of \code{\link[=load]{load()}}
}
