% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_signature_linter.R
\name{local_signature_linter}
\alias{local_signature_linter}
\title{Check signature of \code{\link[=local]{local()}}}
\usage{
local_signature_linter()
}
\description{
Check signature of \code{\link[=local]{local()}}
}
