% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockBinding_signature_linter.R
\name{lockBinding_signature_linter}
\alias{lockBinding_signature_linter}
\title{Check signature of \code{\link[=lockBinding]{lockBinding()}}}
\usage{
lockBinding_signature_linter()
}
\description{
Check signature of \code{\link[=lockBinding]{lockBinding()}}
}
