% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockEnvironment_signature_linter.R
\name{lockEnvironment_signature_linter}
\alias{lockEnvironment_signature_linter}
\title{Check signature of \code{\link[=lockEnvironment]{lockEnvironment()}}}
\usage{
lockEnvironment_signature_linter()
}
\description{
Check signature of \code{\link[=lockEnvironment]{lockEnvironment()}}
}
