% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_signature_linter.R
\name{log_signature_linter}
\alias{log_signature_linter}
\title{Check signature of \code{\link[=log]{log()}}}
\usage{
log_signature_linter()
}
\description{
Check signature of \code{\link[=log]{log()}}
}
