% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logb_signature_linter.R
\name{logb_signature_linter}
\alias{logb_signature_linter}
\title{Check signature of \code{\link[=logb]{logb()}}}
\usage{
logb_signature_linter()
}
\description{
Check signature of \code{\link[=logb]{logb()}}
}
