% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_signature_linter.R
\name{logical_signature_linter}
\alias{logical_signature_linter}
\title{Check signature of \code{\link[=logical]{logical()}}}
\usage{
logical_signature_linter()
}
\description{
Check signature of \code{\link[=logical]{logical()}}
}
