% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_tri_signature_linter.R
\name{lower_tri_signature_linter}
\alias{lower_tri_signature_linter}
\title{Check signature of \code{\link[=lower.tri]{lower.tri()}}}
\usage{
lower_tri_signature_linter()
}
\description{
Check signature of \code{\link[=lower.tri]{lower.tri()}}
}
