% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_names_signature_linter.R
\name{make_names_signature_linter}
\alias{make_names_signature_linter}
\title{Check signature of \code{\link[=make.names]{make.names()}}}
\usage{
make_names_signature_linter()
}
\description{
Check signature of \code{\link[=make.names]{make.names()}}
}
