% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unique_signature_linter.R
\name{make_unique_signature_linter}
\alias{make_unique_signature_linter}
\title{Check signature of \code{\link[=make.unique]{make.unique()}}}
\usage{
make_unique_signature_linter()
}
\description{
Check signature of \code{\link[=make.unique]{make.unique()}}
}
