% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeActiveBinding_signature_linter.R
\name{makeActiveBinding_signature_linter}
\alias{makeActiveBinding_signature_linter}
\title{Check signature of \code{\link[=makeActiveBinding]{makeActiveBinding()}}}
\usage{
makeActiveBinding_signature_linter()
}
\description{
Check signature of \code{\link[=makeActiveBinding]{makeActiveBinding()}}
}
