% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_table_signature_linter.R
\name{margin_table_signature_linter}
\alias{margin_table_signature_linter}
\title{Check signature of \code{\link[=margin.table]{margin.table()}}}
\usage{
margin_table_signature_linter()
}
\description{
Check signature of \code{\link[=margin.table]{margin.table()}}
}
