% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginSums_signature_linter.R
\name{marginSums_signature_linter}
\alias{marginSums_signature_linter}
\title{Check signature of \code{\link[=marginSums]{marginSums()}}}
\usage{
marginSums_signature_linter()
}
\description{
Check signature of \code{\link[=marginSums]{marginSums()}}
}
