% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_or_vec_signature_linter.R
\name{mat_or_vec_signature_linter}
\alias{mat_or_vec_signature_linter}
\title{Check signature of \code{\link[=mat.or.vec]{mat.or.vec()}}}
\usage{
mat_or_vec_signature_linter()
}
\description{
Check signature of \code{\link[=mat.or.vec]{mat.or.vec()}}
}
