% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_arg_signature_linter.R
\name{match_arg_signature_linter}
\alias{match_arg_signature_linter}
\title{Check signature of \code{\link[=match.arg]{match.arg()}}}
\usage{
match_arg_signature_linter()
}
\description{
Check signature of \code{\link[=match.arg]{match.arg()}}
}
