% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_call_signature_linter.R
\name{match_call_signature_linter}
\alias{match_call_signature_linter}
\title{Check signature of \code{\link[=match.call]{match.call()}}}
\usage{
match_call_signature_linter()
}
\description{
Check signature of \code{\link[=match.call]{match.call()}}
}
