% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_fun_signature_linter.R
\name{match_fun_signature_linter}
\alias{match_fun_signature_linter}
\title{Check signature of \code{\link[=match.fun]{match.fun()}}}
\usage{
match_fun_signature_linter()
}
\description{
Check signature of \code{\link[=match.fun]{match.fun()}}
}
