% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math_Date_signature_linter.R
\name{Math_Date_signature_linter}
\alias{Math_Date_signature_linter}
\title{Check signature of \code{\link[=Math.Date]{Math.Date()}}}
\usage{
Math_Date_signature_linter()
}
\description{
Check signature of \code{\link[=Math.Date]{Math.Date()}}
}
