% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math_difftime_signature_linter.R
\name{Math_difftime_signature_linter}
\alias{Math_difftime_signature_linter}
\title{Check signature of \code{\link[=Math.difftime]{Math.difftime()}}}
\usage{
Math_difftime_signature_linter()
}
\description{
Check signature of \code{\link[=Math.difftime]{Math.difftime()}}
}
