% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math_factor_signature_linter.R
\name{Math_factor_signature_linter}
\alias{Math_factor_signature_linter}
\title{Check signature of \code{\link[=Math.factor]{Math.factor()}}}
\usage{
Math_factor_signature_linter()
}
\description{
Check signature of \code{\link[=Math.factor]{Math.factor()}}
}
