% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math_POSIXt_signature_linter.R
\name{Math_POSIXt_signature_linter}
\alias{Math_POSIXt_signature_linter}
\title{Check signature of \code{\link[=Math.POSIXt]{Math.POSIXt()}}}
\usage{
Math_POSIXt_signature_linter()
}
\description{
Check signature of \code{\link[=Math.POSIXt]{Math.POSIXt()}}
}
