% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_signature_linter.R
\name{matrix_signature_linter}
\alias{matrix_signature_linter}
\title{Check signature of \code{\link[=matrix]{matrix()}}}
\usage{
matrix_signature_linter()
}
\description{
Check signature of \code{\link[=matrix]{matrix()}}
}
