% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_col_signature_linter.R
\name{max_col_signature_linter}
\alias{max_col_signature_linter}
\title{Check signature of \code{\link[=max.col]{max.col()}}}
\usage{
max_col_signature_linter()
}
\description{
Check signature of \code{\link[=max.col]{max.col()}}
}
