% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem_maxNSize_signature_linter.R
\name{mem_maxNSize_signature_linter}
\alias{mem_maxNSize_signature_linter}
\title{Check signature of \code{\link[=mem.maxNSize]{mem.maxNSize()}}}
\usage{
mem_maxNSize_signature_linter()
}
\description{
Check signature of \code{\link[=mem.maxNSize]{mem.maxNSize()}}
}
