% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem_maxVSize_signature_linter.R
\name{mem_maxVSize_signature_linter}
\alias{mem_maxVSize_signature_linter}
\title{Check signature of \code{\link[=mem.maxVSize]{mem.maxVSize()}}}
\usage{
mem_maxVSize_signature_linter()
}
\description{
Check signature of \code{\link[=mem.maxVSize]{mem.maxVSize()}}
}
