% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memCompress_signature_linter.R
\name{memCompress_signature_linter}
\alias{memCompress_signature_linter}
\title{Check signature of \code{\link[=memCompress]{memCompress()}}}
\usage{
memCompress_signature_linter()
}
\description{
Check signature of \code{\link[=memCompress]{memCompress()}}
}
