% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mget_signature_linter.R
\name{mget_signature_linter}
\alias{mget_signature_linter}
\title{Check signature of \code{\link[=mget]{mget()}}}
\usage{
mget_signature_linter()
}
\description{
Check signature of \code{\link[=mget]{mget()}}
}
