% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_signature_linter.R
\name{Mod_signature_linter}
\alias{Mod_signature_linter}
\title{Check signature of \code{\link[=Mod]{Mod()}}}
\usage{
Mod_signature_linter()
}
\description{
Check signature of \code{\link[=Mod]{Mod()}}
}
