% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_signature_linter.R
\name{mode_signature_linter}
\alias{mode_signature_linter}
\title{Check signature of \code{\link[=mode]{mode()}}}
\usage{
mode_signature_linter()
}
\description{
Check signature of \code{\link[=mode]{mode()}}
}
