% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/months_POSIXt_signature_linter.R
\name{months_POSIXt_signature_linter}
\alias{months_POSIXt_signature_linter}
\title{Check signature of \code{\link[=months.POSIXt]{months.POSIXt()}}}
\usage{
months_POSIXt_signature_linter()
}
\description{
Check signature of \code{\link[=months.POSIXt]{months.POSIXt()}}
}
