% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/months_signature_linter.R
\name{months_signature_linter}
\alias{months_signature_linter}
\title{Check signature of \code{\link[=months]{months()}}}
\usage{
months_signature_linter()
}
\description{
Check signature of \code{\link[=months]{months()}}
}
