% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtfrm_default_signature_linter.R
\name{mtfrm_default_signature_linter}
\alias{mtfrm_default_signature_linter}
\title{Check signature of \code{\link[=mtfrm.default]{mtfrm.default()}}}
\usage{
mtfrm_default_signature_linter()
}
\description{
Check signature of \code{\link[=mtfrm.default]{mtfrm.default()}}
}
