% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtfrm_POSIXct_signature_linter.R
\name{mtfrm_POSIXct_signature_linter}
\alias{mtfrm_POSIXct_signature_linter}
\title{Check signature of \code{\link[=mtfrm.POSIXct]{mtfrm.POSIXct()}}}
\usage{
mtfrm_POSIXct_signature_linter()
}
\description{
Check signature of \code{\link[=mtfrm.POSIXct]{mtfrm.POSIXct()}}
}
