% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtfrm_POSIXlt_signature_linter.R
\name{mtfrm_POSIXlt_signature_linter}
\alias{mtfrm_POSIXlt_signature_linter}
\title{Check signature of \code{\link[=mtfrm.POSIXlt]{mtfrm.POSIXlt()}}}
\usage{
mtfrm_POSIXlt_signature_linter()
}
\description{
Check signature of \code{\link[=mtfrm.POSIXlt]{mtfrm.POSIXlt()}}
}
