% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtfrm_signature_linter.R
\name{mtfrm_signature_linter}
\alias{mtfrm_signature_linter}
\title{Check signature of \code{\link[=mtfrm]{mtfrm()}}}
\usage{
mtfrm_signature_linter()
}
\description{
Check signature of \code{\link[=mtfrm]{mtfrm()}}
}
