% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameOfClass_default_signature_linter.R
\name{nameOfClass_default_signature_linter}
\alias{nameOfClass_default_signature_linter}
\title{Check signature of \code{\link[=nameOfClass.default]{nameOfClass.default()}}}
\usage{
nameOfClass_default_signature_linter()
}
\description{
Check signature of \code{\link[=nameOfClass.default]{nameOfClass.default()}}
}
