% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameOfClass_signature_linter.R
\name{nameOfClass_signature_linter}
\alias{nameOfClass_signature_linter}
\title{Check signature of \code{\link[=nameOfClass]{nameOfClass()}}}
\usage{
nameOfClass_signature_linter()
}
\description{
Check signature of \code{\link[=nameOfClass]{nameOfClass()}}
}
