% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_signature_linter.R
\name{names_signature_linter}
\alias{names_signature_linter}
\title{Check signature of \code{\link[=names]{names()}}}
\usage{
names_signature_linter()
}
\description{
Check signature of \code{\link[=names]{names()}}
}
