% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespaceExport_signature_linter.R
\name{namespaceExport_signature_linter}
\alias{namespaceExport_signature_linter}
\title{Check signature of \code{\link[=namespaceExport]{namespaceExport()}}}
\usage{
namespaceExport_signature_linter()
}
\description{
Check signature of \code{\link[=namespaceExport]{namespaceExport()}}
}
