% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespaceImportClasses_signature_linter.R
\name{namespaceImportClasses_signature_linter}
\alias{namespaceImportClasses_signature_linter}
\title{Check signature of \code{\link[=namespaceImportClasses]{namespaceImportClasses()}}}
\usage{
namespaceImportClasses_signature_linter()
}
\description{
Check signature of \code{\link[=namespaceImportClasses]{namespaceImportClasses()}}
}
