% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespaceImportFrom_signature_linter.R
\name{namespaceImportFrom_signature_linter}
\alias{namespaceImportFrom_signature_linter}
\title{Check signature of \code{\link[=namespaceImportFrom]{namespaceImportFrom()}}}
\usage{
namespaceImportFrom_signature_linter()
}
\description{
Check signature of \code{\link[=namespaceImportFrom]{namespaceImportFrom()}}
}
