% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespaceImportMethods_signature_linter.R
\name{namespaceImportMethods_signature_linter}
\alias{namespaceImportMethods_signature_linter}
\title{Check signature of \code{\link[=namespaceImportMethods]{namespaceImportMethods()}}}
\usage{
namespaceImportMethods_signature_linter()
}
\description{
Check signature of \code{\link[=namespaceImportMethods]{namespaceImportMethods()}}
}
