% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchar_signature_linter.R
\name{nchar_signature_linter}
\alias{nchar_signature_linter}
\title{Check signature of \code{\link[=nchar]{nchar()}}}
\usage{
nchar_signature_linter()
}
\description{
Check signature of \code{\link[=nchar]{nchar()}}
}
