% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncol_signature_linter.R
\name{ncol_signature_linter}
\alias{ncol_signature_linter}
\title{Check signature of \code{\link[=ncol]{ncol()}}}
\usage{
ncol_signature_linter()
}
\description{
Check signature of \code{\link[=ncol]{ncol()}}
}
