% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Negate_signature_linter.R
\name{Negate_signature_linter}
\alias{Negate_signature_linter}
\title{Check signature of \code{\link[=Negate]{Negate()}}}
\usage{
Negate_signature_linter()
}
\description{
Check signature of \code{\link[=Negate]{Negate()}}
}
