% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_env_signature_linter.R
\name{new_env_signature_linter}
\alias{new_env_signature_linter}
\title{Check signature of \code{\link[=new.env]{new.env()}}}
\usage{
new_env_signature_linter()
}
\description{
Check signature of \code{\link[=new.env]{new.env()}}
}
