% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngettext_signature_linter.R
\name{ngettext_signature_linter}
\alias{ngettext_signature_linter}
\title{Check signature of \code{\link[=ngettext]{ngettext()}}}
\usage{
ngettext_signature_linter()
}
\description{
Check signature of \code{\link[=ngettext]{ngettext()}}
}
