% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlevels_signature_linter.R
\name{nlevels_signature_linter}
\alias{nlevels_signature_linter}
\title{Check signature of \code{\link[=nlevels]{nlevels()}}}
\usage{
nlevels_signature_linter()
}
\description{
Check signature of \code{\link[=nlevels]{nlevels()}}
}
