% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_signature_linter.R
\name{norm_signature_linter}
\alias{norm_signature_linter}
\title{Check signature of \code{\link[=norm]{norm()}}}
\usage{
norm_signature_linter()
}
\description{
Check signature of \code{\link[=norm]{norm()}}
}
