% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizePath_signature_linter.R
\name{normalizePath_signature_linter}
\alias{normalizePath_signature_linter}
\title{Check signature of \code{\link[=normalizePath]{normalizePath()}}}
\usage{
normalizePath_signature_linter()
}
\description{
Check signature of \code{\link[=normalizePath]{normalizePath()}}
}
