% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NROW_signature_linter.R
\name{NROW_signature_linter}
\alias{NROW_signature_linter}
\title{Check signature of \code{\link[=NROW]{NROW()}}}
\usage{
NROW_signature_linter()
}
\description{
Check signature of \code{\link[=NROW]{NROW()}}
}
