% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_signature_linter.R
\name{numeric_signature_linter}
\alias{numeric_signature_linter}
\title{Check signature of \code{\link[=numeric]{numeric()}}}
\usage{
numeric_signature_linter()
}
\description{
Check signature of \code{\link[=numeric]{numeric()}}
}
