% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_version_signature_linter.R
\name{numeric_version_signature_linter}
\alias{numeric_version_signature_linter}
\title{Check signature of \code{\link[=numeric_version]{numeric_version()}}}
\usage{
numeric_version_signature_linter()
}
\description{
Check signature of \code{\link[=numeric_version]{numeric_version()}}
}
