% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numToBits_signature_linter.R
\name{numToBits_signature_linter}
\alias{numToBits_signature_linter}
\title{Check signature of \code{\link[=numToBits]{numToBits()}}}
\usage{
numToBits_signature_linter()
}
\description{
Check signature of \code{\link[=numToBits]{numToBits()}}
}
