% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numToInts_signature_linter.R
\name{numToInts_signature_linter}
\alias{numToInts_signature_linter}
\title{Check signature of \code{\link[=numToInts]{numToInts()}}}
\usage{
numToInts_signature_linter()
}
\description{
Check signature of \code{\link[=numToInts]{numToInts()}}
}
