% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzchar_signature_linter.R
\name{nzchar_signature_linter}
\alias{nzchar_signature_linter}
\title{Check signature of \code{\link[=nzchar]{nzchar()}}}
\usage{
nzchar_signature_linter()
}
\description{
Check signature of \code{\link[=nzchar]{nzchar()}}
}
