% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_signature_linter.R
\name{objects_signature_linter}
\alias{objects_signature_linter}
\title{Check signature of \code{\link[=objects]{objects()}}}
\usage{
objects_signature_linter()
}
\description{
Check signature of \code{\link[=objects]{objects()}}
}
