% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldClass_signature_linter.R
\name{oldClass_signature_linter}
\alias{oldClass_signature_linter}
\title{Check signature of \code{\link[=oldClass]{oldClass()}}}
\usage{
oldClass_signature_linter()
}
\description{
Check signature of \code{\link[=oldClass]{oldClass()}}
}
