% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OlsonNames_signature_linter.R
\name{OlsonNames_signature_linter}
\alias{OlsonNames_signature_linter}
\title{Check signature of \code{\link[=OlsonNames]{OlsonNames()}}}
\usage{
OlsonNames_signature_linter()
}
\description{
Check signature of \code{\link[=OlsonNames]{OlsonNames()}}
}
