% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_exit_signature_linter.R
\name{on_exit_signature_linter}
\alias{on_exit_signature_linter}
\title{Check signature of \code{\link[=on.exit]{on.exit()}}}
\usage{
on_exit_signature_linter()
}
\description{
Check signature of \code{\link[=on.exit]{on.exit()}}
}
