% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_srcfile_signature_linter.R
\name{open_srcfile_signature_linter}
\alias{open_srcfile_signature_linter}
\title{Check signature of \code{\link[=open.srcfile]{open.srcfile()}}}
\usage{
open_srcfile_signature_linter()
}
\description{
Check signature of \code{\link[=open.srcfile]{open.srcfile()}}
}
