% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_srcfilecopy_signature_linter.R
\name{open_srcfilecopy_signature_linter}
\alias{open_srcfilecopy_signature_linter}
\title{Check signature of \code{\link[=open.srcfilecopy]{open.srcfilecopy()}}}
\usage{
open_srcfilecopy_signature_linter()
}
\description{
Check signature of \code{\link[=open.srcfilecopy]{open.srcfilecopy()}}
}
