% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops_data_frame_signature_linter.R
\name{Ops_data_frame_signature_linter}
\alias{Ops_data_frame_signature_linter}
\title{Check signature of \code{\link[=Ops.data.frame]{Ops.data.frame()}}}
\usage{
Ops_data_frame_signature_linter()
}
\description{
Check signature of \code{\link[=Ops.data.frame]{Ops.data.frame()}}
}
