% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops_Date_signature_linter.R
\name{Ops_Date_signature_linter}
\alias{Ops_Date_signature_linter}
\title{Check signature of \code{\link[=Ops.Date]{Ops.Date()}}}
\usage{
Ops_Date_signature_linter()
}
\description{
Check signature of \code{\link[=Ops.Date]{Ops.Date()}}
}
