% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops_difftime_signature_linter.R
\name{Ops_difftime_signature_linter}
\alias{Ops_difftime_signature_linter}
\title{Check signature of \code{\link[=Ops.difftime]{Ops.difftime()}}}
\usage{
Ops_difftime_signature_linter()
}
\description{
Check signature of \code{\link[=Ops.difftime]{Ops.difftime()}}
}
