% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops_factor_signature_linter.R
\name{Ops_factor_signature_linter}
\alias{Ops_factor_signature_linter}
\title{Check signature of \code{\link[=Ops.factor]{Ops.factor()}}}
\usage{
Ops_factor_signature_linter()
}
\description{
Check signature of \code{\link[=Ops.factor]{Ops.factor()}}
}
