% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops_numeric_version_signature_linter.R
\name{Ops_numeric_version_signature_linter}
\alias{Ops_numeric_version_signature_linter}
\title{Check signature of \code{\link[=Ops.numeric_version]{Ops.numeric_version()}}}
\usage{
Ops_numeric_version_signature_linter()
}
\description{
Check signature of \code{\link[=Ops.numeric_version]{Ops.numeric_version()}}
}
