% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops_ordered_signature_linter.R
\name{Ops_ordered_signature_linter}
\alias{Ops_ordered_signature_linter}
\title{Check signature of \code{\link[=Ops.ordered]{Ops.ordered()}}}
\usage{
Ops_ordered_signature_linter()
}
\description{
Check signature of \code{\link[=Ops.ordered]{Ops.ordered()}}
}
