% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops_POSIXt_signature_linter.R
\name{Ops_POSIXt_signature_linter}
\alias{Ops_POSIXt_signature_linter}
\title{Check signature of \code{\link[=Ops.POSIXt]{Ops.POSIXt()}}}
\usage{
Ops_POSIXt_signature_linter()
}
\description{
Check signature of \code{\link[=Ops.POSIXt]{Ops.POSIXt()}}
}
