% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordered_signature_linter.R
\name{ordered_signature_linter}
\alias{ordered_signature_linter}
\title{Check signature of \code{\link[=ordered]{ordered()}}}
\usage{
ordered_signature_linter()
}
\description{
Check signature of \code{\link[=ordered]{ordered()}}
}
