% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_version_signature_linter.R
\name{package_version_signature_linter}
\alias{package_version_signature_linter}
\title{Check signature of \code{\link[=package_version]{package_version()}}}
\usage{
package_version_signature_linter()
}
\description{
Check signature of \code{\link[=package_version]{package_version()}}
}
