% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageEvent_signature_linter.R
\name{packageEvent_signature_linter}
\alias{packageEvent_signature_linter}
\title{Check signature of \code{\link[=packageEvent]{packageEvent()}}}
\usage{
packageEvent_signature_linter()
}
\description{
Check signature of \code{\link[=packageEvent]{packageEvent()}}
}
