% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageHasNamespace_signature_linter.R
\name{packageHasNamespace_signature_linter}
\alias{packageHasNamespace_signature_linter}
\title{Check signature of \code{\link[=packageHasNamespace]{packageHasNamespace()}}}
\usage{
packageHasNamespace_signature_linter()
}
\description{
Check signature of \code{\link[=packageHasNamespace]{packageHasNamespace()}}
}
