% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageNotFoundError_signature_linter.R
\name{packageNotFoundError_signature_linter}
\alias{packageNotFoundError_signature_linter}
\title{Check signature of \code{\link[=packageNotFoundError]{packageNotFoundError()}}}
\usage{
packageNotFoundError_signature_linter()
}
\description{
Check signature of \code{\link[=packageNotFoundError]{packageNotFoundError()}}
}
