% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packBits_signature_linter.R
\name{packBits_signature_linter}
\alias{packBits_signature_linter}
\title{Check signature of \code{\link[=packBits]{packBits()}}}
\usage{
packBits_signature_linter()
}
\description{
Check signature of \code{\link[=packBits]{packBits()}}
}
