% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent_env_signature_linter.R
\name{parent_env_signature_linter}
\alias{parent_env_signature_linter}
\title{Check signature of \code{\link[=parent.env]{parent.env()}}}
\usage{
parent_env_signature_linter()
}
\description{
Check signature of \code{\link[=parent.env]{parent.env()}}
}
