% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent_frame_signature_linter.R
\name{parent_frame_signature_linter}
\alias{parent_frame_signature_linter}
\title{Check signature of \code{\link[=parent.frame]{parent.frame()}}}
\usage{
parent_frame_signature_linter()
}
\description{
Check signature of \code{\link[=parent.frame]{parent.frame()}}
}
