% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseNamespaceFile_signature_linter.R
\name{parseNamespaceFile_signature_linter}
\alias{parseNamespaceFile_signature_linter}
\title{Check signature of \code{\link[=parseNamespaceFile]{parseNamespaceFile()}}}
\usage{
parseNamespaceFile_signature_linter()
}
\description{
Check signature of \code{\link[=parseNamespaceFile]{parseNamespaceFile()}}
}
