% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_expand_signature_linter.R
\name{path_expand_signature_linter}
\alias{path_expand_signature_linter}
\title{Check signature of \code{\link[=path.expand]{path.expand()}}}
\usage{
path_expand_signature_linter()
}
\description{
Check signature of \code{\link[=path.expand]{path.expand()}}
}
