% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_package_signature_linter.R
\name{path_package_signature_linter}
\alias{path_package_signature_linter}
\title{Check signature of \code{\link[=path.package]{path.package()}}}
\usage{
path_package_signature_linter()
}
\description{
Check signature of \code{\link[=path.package]{path.package()}}
}
