% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_signature_linter.R
\name{pipe_signature_linter}
\alias{pipe_signature_linter}
\title{Check signature of \code{\link[=pipe]{pipe()}}}
\usage{
pipe_signature_linter()
}
\description{
Check signature of \code{\link[=pipe]{pipe()}}
}
