% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmatch_signature_linter.R
\name{pmatch_signature_linter}
\alias{pmatch_signature_linter}
\title{Check signature of \code{\link[=pmatch]{pmatch()}}}
\usage{
pmatch_signature_linter()
}
\description{
Check signature of \code{\link[=pmatch]{pmatch()}}
}
