% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyroot_signature_linter.R
\name{polyroot_signature_linter}
\alias{polyroot_signature_linter}
\title{Check signature of \code{\link[=polyroot]{polyroot()}}}
\usage{
polyroot_signature_linter()
}
\description{
Check signature of \code{\link[=polyroot]{polyroot()}}
}
