% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Position_signature_linter.R
\name{Position_signature_linter}
\alias{Position_signature_linter}
\title{Check signature of \code{\link[=Position]{Position()}}}
\usage{
Position_signature_linter()
}
\description{
Check signature of \code{\link[=Position]{Position()}}
}
