% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_AsIs_signature_linter.R
\name{print_AsIs_signature_linter}
\alias{print_AsIs_signature_linter}
\title{Check signature of \code{\link[=print.AsIs]{print.AsIs()}}}
\usage{
print_AsIs_signature_linter()
}
\description{
Check signature of \code{\link[=print.AsIs]{print.AsIs()}}
}
