% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_Dlist_signature_linter.R
\name{print_Dlist_signature_linter}
\alias{print_Dlist_signature_linter}
\title{Check signature of \code{\link[=print.Dlist]{print.Dlist()}}}
\usage{
print_Dlist_signature_linter()
}
\description{
Check signature of \code{\link[=print.Dlist]{print.Dlist()}}
}
