% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_packageInfo_signature_linter.R
\name{print_packageInfo_signature_linter}
\alias{print_packageInfo_signature_linter}
\title{Check signature of \code{\link[=print.packageInfo]{print.packageInfo()}}}
\usage{
print_packageInfo_signature_linter()
}
\description{
Check signature of \code{\link[=print.packageInfo]{print.packageInfo()}}
}
