% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_restart_signature_linter.R
\name{print_restart_signature_linter}
\alias{print_restart_signature_linter}
\title{Check signature of \code{\link[=print.restart]{print.restart()}}}
\usage{
print_restart_signature_linter()
}
\description{
Check signature of \code{\link[=print.restart]{print.restart()}}
}
