% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_rle_signature_linter.R
\name{print_rle_signature_linter}
\alias{print_rle_signature_linter}
\title{Check signature of \code{\link[=print.rle]{print.rle()}}}
\usage{
print_rle_signature_linter()
}
\description{
Check signature of \code{\link[=print.rle]{print.rle()}}
}
