% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_srcfile_signature_linter.R
\name{print_srcfile_signature_linter}
\alias{print_srcfile_signature_linter}
\title{Check signature of \code{\link[=print.srcfile]{print.srcfile()}}}
\usage{
print_srcfile_signature_linter()
}
\description{
Check signature of \code{\link[=print.srcfile]{print.srcfile()}}
}
