% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_srcref_signature_linter.R
\name{print_srcref_signature_linter}
\alias{print_srcref_signature_linter}
\title{Check signature of \code{\link[=print.srcref]{print.srcref()}}}
\usage{
print_srcref_signature_linter()
}
\description{
Check signature of \code{\link[=print.srcref]{print.srcref()}}
}
