% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary_table_signature_linter.R
\name{print_summary_table_signature_linter}
\alias{print_summary_table_signature_linter}
\title{Check signature of \code{\link[=print.summary.table]{print.summary.table()}}}
\usage{
print_summary_table_signature_linter()
}
\description{
Check signature of \code{\link[=print.summary.table]{print.summary.table()}}
}
