% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary_warnings_signature_linter.R
\name{print_summary_warnings_signature_linter}
\alias{print_summary_warnings_signature_linter}
\title{Check signature of \code{\link[=print.summary.warnings]{print.summary.warnings()}}}
\usage{
print_summary_warnings_signature_linter()
}
\description{
Check signature of \code{\link[=print.summary.warnings]{print.summary.warnings()}}
}
