% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_warnings_signature_linter.R
\name{print_warnings_signature_linter}
\alias{print_warnings_signature_linter}
\title{Check signature of \code{\link[=print.warnings]{print.warnings()}}}
\usage{
print_warnings_signature_linter()
}
\description{
Check signature of \code{\link[=print.warnings]{print.warnings()}}
}
