% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_table_signature_linter.R
\name{prop_table_signature_linter}
\alias{prop_table_signature_linter}
\title{Check signature of \code{\link[=prop.table]{prop.table()}}}
\usage{
prop_table_signature_linter()
}
\description{
Check signature of \code{\link[=prop.table]{prop.table()}}
}
