% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions_signature_linter.R
\name{proportions_signature_linter}
\alias{proportions_signature_linter}
\title{Check signature of \code{\link[=proportions]{proportions()}}}
\usage{
proportions_signature_linter()
}
\description{
Check signature of \code{\link[=proportions]{proportions()}}
}
