% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provideDimnames_signature_linter.R
\name{provideDimnames_signature_linter}
\alias{provideDimnames_signature_linter}
\title{Check signature of \code{\link[=provideDimnames]{provideDimnames()}}}
\usage{
provideDimnames_signature_linter()
}
\description{
Check signature of \code{\link[=provideDimnames]{provideDimnames()}}
}
