% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psigamma_signature_linter.R
\name{psigamma_signature_linter}
\alias{psigamma_signature_linter}
\title{Check signature of \code{\link[=psigamma]{psigamma()}}}
\usage{
psigamma_signature_linter()
}
\description{
Check signature of \code{\link[=psigamma]{psigamma()}}
}
