% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushBack_signature_linter.R
\name{pushBack_signature_linter}
\alias{pushBack_signature_linter}
\title{Check signature of \code{\link[=pushBack]{pushBack()}}}
\usage{
pushBack_signature_linter()
}
\description{
Check signature of \code{\link[=pushBack]{pushBack()}}
}
