% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushBackLength_signature_linter.R
\name{pushBackLength_signature_linter}
\alias{pushBackLength_signature_linter}
\title{Check signature of \code{\link[=pushBackLength]{pushBackLength()}}}
\usage{
pushBackLength_signature_linter()
}
\description{
Check signature of \code{\link[=pushBackLength]{pushBackLength()}}
}
