% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_coef_signature_linter.R
\name{qr_coef_signature_linter}
\alias{qr_coef_signature_linter}
\title{Check signature of \code{\link[=qr.coef]{qr.coef()}}}
\usage{
qr_coef_signature_linter()
}
\description{
Check signature of \code{\link[=qr.coef]{qr.coef()}}
}
