% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_fitted_signature_linter.R
\name{qr_fitted_signature_linter}
\alias{qr_fitted_signature_linter}
\title{Check signature of \code{\link[=qr.fitted]{qr.fitted()}}}
\usage{
qr_fitted_signature_linter()
}
\description{
Check signature of \code{\link[=qr.fitted]{qr.fitted()}}
}
