% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_Q_signature_linter.R
\name{qr_Q_signature_linter}
\alias{qr_Q_signature_linter}
\title{Check signature of \code{\link[=qr.Q]{qr.Q()}}}
\usage{
qr_Q_signature_linter()
}
\description{
Check signature of \code{\link[=qr.Q]{qr.Q()}}
}
