% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_qy_signature_linter.R
\name{qr_qy_signature_linter}
\alias{qr_qy_signature_linter}
\title{Check signature of \code{\link[=qr.qy]{qr.qy()}}}
\usage{
qr_qy_signature_linter()
}
\description{
Check signature of \code{\link[=qr.qy]{qr.qy()}}
}
