% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_R_signature_linter.R
\name{qr_R_signature_linter}
\alias{qr_R_signature_linter}
\title{Check signature of \code{\link[=qr.R]{qr.R()}}}
\usage{
qr_R_signature_linter()
}
\description{
Check signature of \code{\link[=qr.R]{qr.R()}}
}
