% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_resid_signature_linter.R
\name{qr_resid_signature_linter}
\alias{qr_resid_signature_linter}
\title{Check signature of \code{\link[=qr.resid]{qr.resid()}}}
\usage{
qr_resid_signature_linter()
}
\description{
Check signature of \code{\link[=qr.resid]{qr.resid()}}
}
