% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_X_signature_linter.R
\name{qr_X_signature_linter}
\alias{qr_X_signature_linter}
\title{Check signature of \code{\link[=qr.X]{qr.X()}}}
\usage{
qr_X_signature_linter()
}
\description{
Check signature of \code{\link[=qr.X]{qr.X()}}
}
