% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarters_Date_signature_linter.R
\name{quarters_Date_signature_linter}
\alias{quarters_Date_signature_linter}
\title{Check signature of \code{\link[=quarters.Date]{quarters.Date()}}}
\usage{
quarters_Date_signature_linter()
}
\description{
Check signature of \code{\link[=quarters.Date]{quarters.Date()}}
}
