% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarters_POSIXt_signature_linter.R
\name{quarters_POSIXt_signature_linter}
\alias{quarters_POSIXt_signature_linter}
\title{Check signature of \code{\link[=quarters.POSIXt]{quarters.POSIXt()}}}
\usage{
quarters_POSIXt_signature_linter()
}
\description{
Check signature of \code{\link[=quarters.POSIXt]{quarters.POSIXt()}}
}
