% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarters_signature_linter.R
\name{quarters_signature_linter}
\alias{quarters_signature_linter}
\title{Check signature of \code{\link[=quarters]{quarters()}}}
\usage{
quarters_signature_linter()
}
\description{
Check signature of \code{\link[=quarters]{quarters()}}
}
