% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quit_signature_linter.R
\name{quit_signature_linter}
\alias{quit_signature_linter}
\title{Check signature of \code{\link[=quit]{quit()}}}
\usage{
quit_signature_linter()
}
\description{
Check signature of \code{\link[=quit]{quit()}}
}
