% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote_signature_linter.R
\name{quote_signature_linter}
\alias{quote_signature_linter}
\title{Check signature of \code{\link[=quote]{quote()}}}
\usage{
quote_signature_linter()
}
\description{
Check signature of \code{\link[=quote]{quote()}}
}
