% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_home_signature_linter.R
\name{R_home_signature_linter}
\alias{R_home_signature_linter}
\title{Check signature of \code{\link[=R.home]{R.home()}}}
\usage{
R_home_signature_linter()
}
\description{
Check signature of \code{\link[=R.home]{R.home()}}
}
