% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_system_version_signature_linter.R
\name{R_system_version_signature_linter}
\alias{R_system_version_signature_linter}
\title{Check signature of \code{\link[=R_system_version]{R_system_version()}}}
\usage{
R_system_version_signature_linter()
}
\description{
Check signature of \code{\link[=R_system_version]{R_system_version()}}
}
