% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_signature_linter.R
\name{rank_signature_linter}
\alias{rank_signature_linter}
\title{Check signature of \code{\link[=rank]{rank()}}}
\usage{
rank_signature_linter()
}
\description{
Check signature of \code{\link[=rank]{rank()}}
}
