% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_signature_linter.R
\name{raw_signature_linter}
\alias{raw_signature_linter}
\title{Check signature of \code{\link[=raw]{raw()}}}
\usage{
raw_signature_linter()
}
\description{
Check signature of \code{\link[=raw]{raw()}}
}
