% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawConnection_signature_linter.R
\name{rawConnection_signature_linter}
\alias{rawConnection_signature_linter}
\title{Check signature of \code{\link[=rawConnection]{rawConnection()}}}
\usage{
rawConnection_signature_linter()
}
\description{
Check signature of \code{\link[=rawConnection]{rawConnection()}}
}
