% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawConnectionValue_signature_linter.R
\name{rawConnectionValue_signature_linter}
\alias{rawConnectionValue_signature_linter}
\title{Check signature of \code{\link[=rawConnectionValue]{rawConnectionValue()}}}
\usage{
rawConnectionValue_signature_linter()
}
\description{
Check signature of \code{\link[=rawConnectionValue]{rawConnectionValue()}}
}
