% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawShift_signature_linter.R
\name{rawShift_signature_linter}
\alias{rawShift_signature_linter}
\title{Check signature of \code{\link[=rawShift]{rawShift()}}}
\usage{
rawShift_signature_linter()
}
\description{
Check signature of \code{\link[=rawShift]{rawShift()}}
}
