% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawToBits_signature_linter.R
\name{rawToBits_signature_linter}
\alias{rawToBits_signature_linter}
\title{Check signature of \code{\link[=rawToBits]{rawToBits()}}}
\usage{
rawToBits_signature_linter()
}
\description{
Check signature of \code{\link[=rawToBits]{rawToBits()}}
}
