% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dcf_signature_linter.R
\name{read_dcf_signature_linter}
\alias{read_dcf_signature_linter}
\title{Check signature of \code{\link[=read.dcf]{read.dcf()}}}
\usage{
read_dcf_signature_linter()
}
\description{
Check signature of \code{\link[=read.dcf]{read.dcf()}}
}
