% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBin_signature_linter.R
\name{readBin_signature_linter}
\alias{readBin_signature_linter}
\title{Check signature of \code{\link[=readBin]{readBin()}}}
\usage{
readBin_signature_linter()
}
\description{
Check signature of \code{\link[=readBin]{readBin()}}
}
