% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readChar_signature_linter.R
\name{readChar_signature_linter}
\alias{readChar_signature_linter}
\title{Check signature of \code{\link[=readChar]{readChar()}}}
\usage{
readChar_signature_linter()
}
\description{
Check signature of \code{\link[=readChar]{readChar()}}
}
