% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readline_signature_linter.R
\name{readline_signature_linter}
\alias{readline_signature_linter}
\title{Check signature of \code{\link[=readline]{readline()}}}
\usage{
readline_signature_linter()
}
\description{
Check signature of \code{\link[=readline]{readline()}}
}
