% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLines_signature_linter.R
\name{readLines_signature_linter}
\alias{readLines_signature_linter}
\title{Check signature of \code{\link[=readLines]{readLines()}}}
\usage{
readLines_signature_linter()
}
\description{
Check signature of \code{\link[=readLines]{readLines()}}
}
