% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRDS_signature_linter.R
\name{readRDS_signature_linter}
\alias{readRDS_signature_linter}
\title{Check signature of \code{\link[=readRDS]{readRDS()}}}
\usage{
readRDS_signature_linter()
}
\description{
Check signature of \code{\link[=readRDS]{readRDS()}}
}
