% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRenviron_signature_linter.R
\name{readRenviron_signature_linter}
\alias{readRenviron_signature_linter}
\title{Check signature of \code{\link[=readRenviron]{readRenviron()}}}
\usage{
readRenviron_signature_linter()
}
\description{
Check signature of \code{\link[=readRenviron]{readRenviron()}}
}
