% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_signature_linter.R
\name{Reduce_signature_linter}
\alias{Reduce_signature_linter}
\title{Check signature of \code{\link[=Reduce]{Reduce()}}}
\usage{
Reduce_signature_linter()
}
\description{
Check signature of \code{\link[=Reduce]{Reduce()}}
}
