% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_finalizer_signature_linter.R
\name{reg_finalizer_signature_linter}
\alias{reg_finalizer_signature_linter}
\title{Check signature of \code{\link[=reg.finalizer]{reg.finalizer()}}}
\usage{
reg_finalizer_signature_linter()
}
\description{
Check signature of \code{\link[=reg.finalizer]{reg.finalizer()}}
}
