% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexec_signature_linter.R
\name{regexec_signature_linter}
\alias{regexec_signature_linter}
\title{Check signature of \code{\link[=regexec]{regexec()}}}
\usage{
regexec_signature_linter()
}
\description{
Check signature of \code{\link[=regexec]{regexec()}}
}
