% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexpr_signature_linter.R
\name{regexpr_signature_linter}
\alias{regexpr_signature_linter}
\title{Check signature of \code{\link[=regexpr]{regexpr()}}}
\usage{
regexpr_signature_linter()
}
\description{
Check signature of \code{\link[=regexpr]{regexpr()}}
}
