% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmatches_signature_linter.R
\name{regmatches_signature_linter}
\alias{regmatches_signature_linter}
\title{Check signature of \code{\link[=regmatches]{regmatches()}}}
\usage{
regmatches_signature_linter()
}
\description{
Check signature of \code{\link[=regmatches]{regmatches()}}
}
