% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_Date_signature_linter.R
\name{rep_Date_signature_linter}
\alias{rep_Date_signature_linter}
\title{Check signature of \code{\link[=rep.Date]{rep.Date()}}}
\usage{
rep_Date_signature_linter()
}
\description{
Check signature of \code{\link[=rep.Date]{rep.Date()}}
}
