% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_difftime_signature_linter.R
\name{rep_difftime_signature_linter}
\alias{rep_difftime_signature_linter}
\title{Check signature of \code{\link[=rep.difftime]{rep.difftime()}}}
\usage{
rep_difftime_signature_linter()
}
\description{
Check signature of \code{\link[=rep.difftime]{rep.difftime()}}
}
