% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_factor_signature_linter.R
\name{rep_factor_signature_linter}
\alias{rep_factor_signature_linter}
\title{Check signature of \code{\link[=rep.factor]{rep.factor()}}}
\usage{
rep_factor_signature_linter()
}
\description{
Check signature of \code{\link[=rep.factor]{rep.factor()}}
}
