% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_int_signature_linter.R
\name{rep_int_signature_linter}
\alias{rep_int_signature_linter}
\title{Check signature of \code{\link[=rep.int]{rep.int()}}}
\usage{
rep_int_signature_linter()
}
\description{
Check signature of \code{\link[=rep.int]{rep.int()}}
}
