% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_len_signature_linter.R
\name{rep_len_signature_linter}
\alias{rep_len_signature_linter}
\title{Check signature of \code{\link[=rep_len]{rep_len()}}}
\usage{
rep_len_signature_linter()
}
\description{
Check signature of \code{\link[=rep_len]{rep_len()}}
}
