% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_numeric_version_signature_linter.R
\name{rep_numeric_version_signature_linter}
\alias{rep_numeric_version_signature_linter}
\title{Check signature of \code{\link[=rep.numeric_version]{rep.numeric_version()}}}
\usage{
rep_numeric_version_signature_linter()
}
\description{
Check signature of \code{\link[=rep.numeric_version]{rep.numeric_version()}}
}
