% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_POSIXct_signature_linter.R
\name{rep_POSIXct_signature_linter}
\alias{rep_POSIXct_signature_linter}
\title{Check signature of \code{\link[=rep.POSIXct]{rep.POSIXct()}}}
\usage{
rep_POSIXct_signature_linter()
}
\description{
Check signature of \code{\link[=rep.POSIXct]{rep.POSIXct()}}
}
