% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_signature_linter.R
\name{replicate_signature_linter}
\alias{replicate_signature_linter}
\title{Check signature of \code{\link[=replicate]{replicate()}}}
\usage{
replicate_signature_linter()
}
\description{
Check signature of \code{\link[=replicate]{replicate()}}
}
