% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require_signature_linter.R
\name{require_signature_linter}
\alias{require_signature_linter}
\title{Check signature of \code{\link[=require]{require()}}}
\usage{
require_signature_linter()
}
\description{
Check signature of \code{\link[=require]{require()}}
}
