% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restartDescription_signature_linter.R
\name{restartDescription_signature_linter}
\alias{restartDescription_signature_linter}
\title{Check signature of \code{\link[=restartDescription]{restartDescription()}}}
\usage{
restartDescription_signature_linter()
}
\description{
Check signature of \code{\link[=restartDescription]{restartDescription()}}
}
