% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restartFormals_signature_linter.R
\name{restartFormals_signature_linter}
\alias{restartFormals_signature_linter}
\title{Check signature of \code{\link[=restartFormals]{restartFormals()}}}
\usage{
restartFormals_signature_linter()
}
\description{
Check signature of \code{\link[=restartFormals]{restartFormals()}}
}
