% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnValue_signature_linter.R
\name{returnValue_signature_linter}
\alias{returnValue_signature_linter}
\title{Check signature of \code{\link[=returnValue]{returnValue()}}}
\usage{
returnValue_signature_linter()
}
\description{
Check signature of \code{\link[=returnValue]{returnValue()}}
}
