% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_default_signature_linter.R
\name{rev_default_signature_linter}
\alias{rev_default_signature_linter}
\title{Check signature of \code{\link[=rev.default]{rev.default()}}}
\usage{
rev_default_signature_linter()
}
\description{
Check signature of \code{\link[=rev.default]{rev.default()}}
}
