% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_signature_linter.R
\name{rev_signature_linter}
\alias{rev_signature_linter}
\title{Check signature of \code{\link[=rev]{rev()}}}
\usage{
rev_signature_linter()
}
\description{
Check signature of \code{\link[=rev]{rev()}}
}
