% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle_signature_linter.R
\name{rle_signature_linter}
\alias{rle_signature_linter}
\title{Check signature of \code{\link[=rle]{rle()}}}
\usage{
rle_signature_linter()
}
\description{
Check signature of \code{\link[=rle]{rle()}}
}
