% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNGkind_signature_linter.R
\name{RNGkind_signature_linter}
\alias{RNGkind_signature_linter}
\title{Check signature of \code{\link[=RNGkind]{RNGkind()}}}
\usage{
RNGkind_signature_linter()
}
\description{
Check signature of \code{\link[=RNGkind]{RNGkind()}}
}
