% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNGversion_signature_linter.R
\name{RNGversion_signature_linter}
\alias{RNGversion_signature_linter}
\title{Check signature of \code{\link[=RNGversion]{RNGversion()}}}
\usage{
RNGversion_signature_linter()
}
\description{
Check signature of \code{\link[=RNGversion]{RNGversion()}}
}
