% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_Date_signature_linter.R
\name{round_Date_signature_linter}
\alias{round_Date_signature_linter}
\title{Check signature of \code{\link[=round.Date]{round.Date()}}}
\usage{
round_Date_signature_linter()
}
\description{
Check signature of \code{\link[=round.Date]{round.Date()}}
}
