% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_POSIXt_signature_linter.R
\name{round_POSIXt_signature_linter}
\alias{round_POSIXt_signature_linter}
\title{Check signature of \code{\link[=round.POSIXt]{round.POSIXt()}}}
\usage{
round_POSIXt_signature_linter()
}
\description{
Check signature of \code{\link[=round.POSIXt]{round.POSIXt()}}
}
