% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_names_default_signature_linter.R
\name{row_names_default_signature_linter}
\alias{row_names_default_signature_linter}
\title{Check signature of \code{\link[=row.names.default]{row.names.default()}}}
\usage{
row_names_default_signature_linter()
}
\description{
Check signature of \code{\link[=row.names.default]{row.names.default()}}
}
