% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowMeans_signature_linter.R
\name{rowMeans_signature_linter}
\alias{rowMeans_signature_linter}
\title{Check signature of \code{\link[=rowMeans]{rowMeans()}}}
\usage{
rowMeans_signature_linter()
}
\description{
Check signature of \code{\link[=rowMeans]{rowMeans()}}
}
