% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rownames_signature_linter.R
\name{rownames_signature_linter}
\alias{rownames_signature_linter}
\title{Check signature of \code{\link[=rownames]{rownames()}}}
\usage{
rownames_signature_linter()
}
\description{
Check signature of \code{\link[=rownames]{rownames()}}
}
