% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowSums_signature_linter.R
\name{rowSums_signature_linter}
\alias{rowSums_signature_linter}
\title{Check signature of \code{\link[=rowSums]{rowSums()}}}
\usage{
rowSums_signature_linter()
}
\description{
Check signature of \code{\link[=rowSums]{rowSums()}}
}
