% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_signature_linter.R
\name{sample_signature_linter}
\alias{sample_signature_linter}
\title{Check signature of \code{\link[=sample]{sample()}}}
\usage{
sample_signature_linter()
}
\description{
Check signature of \code{\link[=sample]{sample()}}
}
