% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_image_signature_linter.R
\name{save_image_signature_linter}
\alias{save_image_signature_linter}
\title{Check signature of \code{\link[=save.image]{save.image()}}}
\usage{
save_image_signature_linter()
}
\description{
Check signature of \code{\link[=save.image]{save.image()}}
}
