% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRDS_signature_linter.R
\name{saveRDS_signature_linter}
\alias{saveRDS_signature_linter}
\title{Check signature of \code{\link[=saveRDS]{saveRDS()}}}
\usage{
saveRDS_signature_linter()
}
\description{
Check signature of \code{\link[=saveRDS]{saveRDS()}}
}
