% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_signature_linter.R
\name{scale_signature_linter}
\alias{scale_signature_linter}
\title{Check signature of \code{\link[=scale]{scale()}}}
\usage{
scale_signature_linter()
}
\description{
Check signature of \code{\link[=scale]{scale()}}
}
