% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_signature_linter.R
\name{scan_signature_linter}
\alias{scan_signature_linter}
\title{Check signature of \code{\link[=scan]{scan()}}}
\usage{
scan_signature_linter()
}
\description{
Check signature of \code{\link[=scan]{scan()}}
}
