% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_along_signature_linter.R
\name{seq_along_signature_linter}
\alias{seq_along_signature_linter}
\title{Check signature of \code{\link[=seq_along]{seq_along()}}}
\usage{
seq_along_signature_linter()
}
\description{
Check signature of \code{\link[=seq_along]{seq_along()}}
}
