% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_Date_signature_linter.R
\name{seq_Date_signature_linter}
\alias{seq_Date_signature_linter}
\title{Check signature of \code{\link[=seq.Date]{seq.Date()}}}
\usage{
seq_Date_signature_linter()
}
\description{
Check signature of \code{\link[=seq.Date]{seq.Date()}}
}
