% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_default_signature_linter.R
\name{seq_default_signature_linter}
\alias{seq_default_signature_linter}
\title{Check signature of \code{\link[=seq.default]{seq.default()}}}
\usage{
seq_default_signature_linter()
}
\description{
Check signature of \code{\link[=seq.default]{seq.default()}}
}
