% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_len_signature_linter.R
\name{seq_len_signature_linter}
\alias{seq_len_signature_linter}
\title{Check signature of \code{\link[=seq_len]{seq_len()}}}
\usage{
seq_len_signature_linter()
}
\description{
Check signature of \code{\link[=seq_len]{seq_len()}}
}
