% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_POSIXt_signature_linter.R
\name{seq_POSIXt_signature_linter}
\alias{seq_POSIXt_signature_linter}
\title{Check signature of \code{\link[=seq.POSIXt]{seq.POSIXt()}}}
\usage{
seq_POSIXt_signature_linter()
}
\description{
Check signature of \code{\link[=seq.POSIXt]{seq.POSIXt()}}
}
