% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize_signature_linter.R
\name{serialize_signature_linter}
\alias{serialize_signature_linter}
\title{Check signature of \code{\link[=serialize]{serialize()}}}
\usage{
serialize_signature_linter()
}
\description{
Check signature of \code{\link[=serialize]{serialize()}}
}
