% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverSocket_signature_linter.R
\name{serverSocket_signature_linter}
\alias{serverSocket_signature_linter}
\title{Check signature of \code{\link[=serverSocket]{serverSocket()}}}
\usage{
serverSocket_signature_linter()
}
\description{
Check signature of \code{\link[=serverSocket]{serverSocket()}}
}
