% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_seed_signature_linter.R
\name{set_seed_signature_linter}
\alias{set_seed_signature_linter}
\title{Check signature of \code{\link[=set.seed]{set.seed()}}}
\usage{
set_seed_signature_linter()
}
\description{
Check signature of \code{\link[=set.seed]{set.seed()}}
}
