% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdiff_signature_linter.R
\name{setdiff_signature_linter}
\alias{setdiff_signature_linter}
\title{Check signature of \code{\link[=setdiff]{setdiff()}}}
\usage{
setdiff_signature_linter()
}
\description{
Check signature of \code{\link[=setdiff]{setdiff()}}
}
