% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setequal_signature_linter.R
\name{setequal_signature_linter}
\alias{setequal_signature_linter}
\title{Check signature of \code{\link[=setequal]{setequal()}}}
\usage{
setequal_signature_linter()
}
\description{
Check signature of \code{\link[=setequal]{setequal()}}
}
