% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setHook_signature_linter.R
\name{setHook_signature_linter}
\alias{setHook_signature_linter}
\title{Check signature of \code{\link[=setHook]{setHook()}}}
\usage{
setHook_signature_linter()
}
\description{
Check signature of \code{\link[=setHook]{setHook()}}
}
