% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNamespaceInfo_signature_linter.R
\name{setNamespaceInfo_signature_linter}
\alias{setNamespaceInfo_signature_linter}
\title{Check signature of \code{\link[=setNamespaceInfo]{setNamespaceInfo()}}}
\usage{
setNamespaceInfo_signature_linter()
}
\description{
Check signature of \code{\link[=setNamespaceInfo]{setNamespaceInfo()}}
}
