% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSessionTimeLimit_signature_linter.R
\name{setSessionTimeLimit_signature_linter}
\alias{setSessionTimeLimit_signature_linter}
\title{Check signature of \code{\link[=setSessionTimeLimit]{setSessionTimeLimit()}}}
\usage{
setSessionTimeLimit_signature_linter()
}
\description{
Check signature of \code{\link[=setSessionTimeLimit]{setSessionTimeLimit()}}
}
