% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTimeLimit_signature_linter.R
\name{setTimeLimit_signature_linter}
\alias{setTimeLimit_signature_linter}
\title{Check signature of \code{\link[=setTimeLimit]{setTimeLimit()}}}
\usage{
setTimeLimit_signature_linter()
}
\description{
Check signature of \code{\link[=setTimeLimit]{setTimeLimit()}}
}
