% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd_signature_linter.R
\name{setwd_signature_linter}
\alias{setwd_signature_linter}
\title{Check signature of \code{\link[=setwd]{setwd()}}}
\usage{
setwd_signature_linter()
}
\description{
Check signature of \code{\link[=setwd]{setwd()}}
}
