% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showConnections_signature_linter.R
\name{showConnections_signature_linter}
\alias{showConnections_signature_linter}
\title{Check signature of \code{\link[=showConnections]{showConnections()}}}
\usage{
showConnections_signature_linter()
}
\description{
Check signature of \code{\link[=showConnections]{showConnections()}}
}
