% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shQuote_signature_linter.R
\name{shQuote_signature_linter}
\alias{shQuote_signature_linter}
\title{Check signature of \code{\link[=shQuote]{shQuote()}}}
\usage{
shQuote_signature_linter()
}
\description{
Check signature of \code{\link[=shQuote]{shQuote()}}
}
