% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_signature_linter.R
\name{sign_signature_linter}
\alias{sign_signature_linter}
\title{Check signature of \code{\link[=sign]{sign()}}}
\usage{
sign_signature_linter()
}
\description{
Check signature of \code{\link[=sign]{sign()}}
}
