% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalCondition_signature_linter.R
\name{signalCondition_signature_linter}
\alias{signalCondition_signature_linter}
\title{Check signature of \code{\link[=signalCondition]{signalCondition()}}}
\usage{
signalCondition_signature_linter()
}
\description{
Check signature of \code{\link[=signalCondition]{signalCondition()}}
}
