% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_signature_linter.R
\name{signif_signature_linter}
\alias{signif_signature_linter}
\title{Check signature of \code{\link[=signif]{signif()}}}
\usage{
signif_signature_linter()
}
\description{
Check signature of \code{\link[=signif]{signif()}}
}
