% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCondition_signature_linter.R
\name{simpleCondition_signature_linter}
\alias{simpleCondition_signature_linter}
\title{Check signature of \code{\link[=simpleCondition]{simpleCondition()}}}
\usage{
simpleCondition_signature_linter()
}
\description{
Check signature of \code{\link[=simpleCondition]{simpleCondition()}}
}
