% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleError_signature_linter.R
\name{simpleError_signature_linter}
\alias{simpleError_signature_linter}
\title{Check signature of \code{\link[=simpleError]{simpleError()}}}
\usage{
simpleError_signature_linter()
}
\description{
Check signature of \code{\link[=simpleError]{simpleError()}}
}
