% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleMessage_signature_linter.R
\name{simpleMessage_signature_linter}
\alias{simpleMessage_signature_linter}
\title{Check signature of \code{\link[=simpleMessage]{simpleMessage()}}}
\usage{
simpleMessage_signature_linter()
}
\description{
Check signature of \code{\link[=simpleMessage]{simpleMessage()}}
}
