% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleWarning_signature_linter.R
\name{simpleWarning_signature_linter}
\alias{simpleWarning_signature_linter}
\title{Check signature of \code{\link[=simpleWarning]{simpleWarning()}}}
\usage{
simpleWarning_signature_linter()
}
\description{
Check signature of \code{\link[=simpleWarning]{simpleWarning()}}
}
