% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sin_signature_linter.R
\name{sin_signature_linter}
\alias{sin_signature_linter}
\title{Check signature of \code{\link[=sin]{sin()}}}
\usage{
sin_signature_linter()
}
\description{
Check signature of \code{\link[=sin]{sin()}}
}
