% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_signature_linter.R
\name{single_signature_linter}
\alias{single_signature_linter}
\title{Check signature of \code{\link[=single]{single()}}}
\usage{
single_signature_linter()
}
\description{
Check signature of \code{\link[=single]{single()}}
}
