% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinh_signature_linter.R
\name{sinh_signature_linter}
\alias{sinh_signature_linter}
\title{Check signature of \code{\link[=sinh]{sinh()}}}
\usage{
sinh_signature_linter()
}
\description{
Check signature of \code{\link[=sinh]{sinh()}}
}
