% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink_number_signature_linter.R
\name{sink_number_signature_linter}
\alias{sink_number_signature_linter}
\title{Check signature of \code{\link[=sink.number]{sink.number()}}}
\usage{
sink_number_signature_linter()
}
\description{
Check signature of \code{\link[=sink.number]{sink.number()}}
}
