% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink_signature_linter.R
\name{sink_signature_linter}
\alias{sink_signature_linter}
\title{Check signature of \code{\link[=sink]{sink()}}}
\usage{
sink_signature_linter()
}
\description{
Check signature of \code{\link[=sink]{sink()}}
}
