% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinpi_signature_linter.R
\name{sinpi_signature_linter}
\alias{sinpi_signature_linter}
\title{Check signature of \code{\link[=sinpi]{sinpi()}}}
\usage{
sinpi_signature_linter()
}
\description{
Check signature of \code{\link[=sinpi]{sinpi()}}
}
