% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_index_signature_linter.R
\name{slice_index_signature_linter}
\alias{slice_index_signature_linter}
\title{Check signature of \code{\link[=slice.index]{slice.index()}}}
\usage{
slice_index_signature_linter()
}
\description{
Check signature of \code{\link[=slice.index]{slice.index()}}
}
