% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socketAccept_signature_linter.R
\name{socketAccept_signature_linter}
\alias{socketAccept_signature_linter}
\title{Check signature of \code{\link[=socketAccept]{socketAccept()}}}
\usage{
socketAccept_signature_linter()
}
\description{
Check signature of \code{\link[=socketAccept]{socketAccept()}}
}
