% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socketConnection_signature_linter.R
\name{socketConnection_signature_linter}
\alias{socketConnection_signature_linter}
\title{Check signature of \code{\link[=socketConnection]{socketConnection()}}}
\usage{
socketConnection_signature_linter()
}
\description{
Check signature of \code{\link[=socketConnection]{socketConnection()}}
}
