% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socketSelect_signature_linter.R
\name{socketSelect_signature_linter}
\alias{socketSelect_signature_linter}
\title{Check signature of \code{\link[=socketSelect]{socketSelect()}}}
\usage{
socketSelect_signature_linter()
}
\description{
Check signature of \code{\link[=socketSelect]{socketSelect()}}
}
