% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socketTimeout_signature_linter.R
\name{socketTimeout_signature_linter}
\alias{socketTimeout_signature_linter}
\title{Check signature of \code{\link[=socketTimeout]{socketTimeout()}}}
\usage{
socketTimeout_signature_linter()
}
\description{
Check signature of \code{\link[=socketTimeout]{socketTimeout()}}
}
