% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_default_signature_linter.R
\name{sort_default_signature_linter}
\alias{sort_default_signature_linter}
\title{Check signature of \code{\link[=sort.default]{sort.default()}}}
\usage{
sort_default_signature_linter()
}
\description{
Check signature of \code{\link[=sort.default]{sort.default()}}
}
