% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_int_signature_linter.R
\name{sort_int_signature_linter}
\alias{sort_int_signature_linter}
\title{Check signature of \code{\link[=sort.int]{sort.int()}}}
\usage{
sort_int_signature_linter()
}
\description{
Check signature of \code{\link[=sort.int]{sort.int()}}
}
