% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_list_signature_linter.R
\name{sort_list_signature_linter}
\alias{sort_list_signature_linter}
\title{Check signature of \code{\link[=sort.list]{sort.list()}}}
\usage{
sort_list_signature_linter()
}
\description{
Check signature of \code{\link[=sort.list]{sort.list()}}
}
