% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_POSIXlt_signature_linter.R
\name{sort_POSIXlt_signature_linter}
\alias{sort_POSIXlt_signature_linter}
\title{Check signature of \code{\link[=sort.POSIXlt]{sort.POSIXlt()}}}
\usage{
sort_POSIXlt_signature_linter()
}
\description{
Check signature of \code{\link[=sort.POSIXlt]{sort.POSIXlt()}}
}
