% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_signature_linter.R
\name{source_signature_linter}
\alias{source_signature_linter}
\title{Check signature of \code{\link[=source]{source()}}}
\usage{
source_signature_linter()
}
\description{
Check signature of \code{\link[=source]{source()}}
}
