% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrt_signature_linter.R
\name{sqrt_signature_linter}
\alias{sqrt_signature_linter}
\title{Check signature of \code{\link[=sqrt]{sqrt()}}}
\usage{
sqrt_signature_linter()
}
\description{
Check signature of \code{\link[=sqrt]{sqrt()}}
}
