% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sQuote_signature_linter.R
\name{sQuote_signature_linter}
\alias{sQuote_signature_linter}
\title{Check signature of \code{\link[=sQuote]{sQuote()}}}
\usage{
sQuote_signature_linter()
}
\description{
Check signature of \code{\link[=sQuote]{sQuote()}}
}
