% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcfile_signature_linter.R
\name{srcfile_signature_linter}
\alias{srcfile_signature_linter}
\title{Check signature of \code{\link[=srcfile]{srcfile()}}}
\usage{
srcfile_signature_linter()
}
\description{
Check signature of \code{\link[=srcfile]{srcfile()}}
}
