% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcfilealias_signature_linter.R
\name{srcfilealias_signature_linter}
\alias{srcfilealias_signature_linter}
\title{Check signature of \code{\link[=srcfilealias]{srcfilealias()}}}
\usage{
srcfilealias_signature_linter()
}
\description{
Check signature of \code{\link[=srcfilealias]{srcfilealias()}}
}
