% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcfilecopy_signature_linter.R
\name{srcfilecopy_signature_linter}
\alias{srcfilecopy_signature_linter}
\title{Check signature of \code{\link[=srcfilecopy]{srcfilecopy()}}}
\usage{
srcfilecopy_signature_linter()
}
\description{
Check signature of \code{\link[=srcfilecopy]{srcfilecopy()}}
}
