% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcref_signature_linter.R
\name{srcref_signature_linter}
\alias{srcref_signature_linter}
\title{Check signature of \code{\link[=srcref]{srcref()}}}
\usage{
srcref_signature_linter()
}
\description{
Check signature of \code{\link[=srcref]{srcref()}}
}
