% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardGeneric_signature_linter.R
\name{standardGeneric_signature_linter}
\alias{standardGeneric_signature_linter}
\title{Check signature of \code{\link[=standardGeneric]{standardGeneric()}}}
\usage{
standardGeneric_signature_linter()
}
\description{
Check signature of \code{\link[=standardGeneric]{standardGeneric()}}
}
