% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startsWith_signature_linter.R
\name{startsWith_signature_linter}
\alias{startsWith_signature_linter}
\title{Check signature of \code{\link[=startsWith]{startsWith()}}}
\usage{
startsWith_signature_linter()
}
\description{
Check signature of \code{\link[=startsWith]{startsWith()}}
}
