% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopifnot_signature_linter.R
\name{stopifnot_signature_linter}
\alias{stopifnot_signature_linter}
\title{Check signature of \code{\link[=stopifnot]{stopifnot()}}}
\usage{
stopifnot_signature_linter()
}
\description{
Check signature of \code{\link[=stopifnot]{stopifnot()}}
}
