% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_mode_signature_linter.R
\name{storage_mode_signature_linter}
\alias{storage_mode_signature_linter}
\title{Check signature of \code{\link[=storage.mode]{storage.mode()}}}
\usage{
storage_mode_signature_linter()
}
\description{
Check signature of \code{\link[=storage.mode]{storage.mode()}}
}
