% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strptime_signature_linter.R
\name{strptime_signature_linter}
\alias{strptime_signature_linter}
\title{Check signature of \code{\link[=strptime]{strptime()}}}
\usage{
strptime_signature_linter()
}
\description{
Check signature of \code{\link[=strptime]{strptime()}}
}
