% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit_signature_linter.R
\name{strsplit_signature_linter}
\alias{strsplit_signature_linter}
\title{Check signature of \code{\link[=strsplit]{strsplit()}}}
\usage{
strsplit_signature_linter()
}
\description{
Check signature of \code{\link[=strsplit]{strsplit()}}
}
