% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strtoi_signature_linter.R
\name{strtoi_signature_linter}
\alias{strtoi_signature_linter}
\title{Check signature of \code{\link[=strtoi]{strtoi()}}}
\usage{
strtoi_signature_linter()
}
\description{
Check signature of \code{\link[=strtoi]{strtoi()}}
}
