% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strtrim_signature_linter.R
\name{strtrim_signature_linter}
\alias{strtrim_signature_linter}
\title{Check signature of \code{\link[=strtrim]{strtrim()}}}
\usage{
strtrim_signature_linter()
}
\description{
Check signature of \code{\link[=strtrim]{strtrim()}}
}
