% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strwrap_signature_linter.R
\name{strwrap_signature_linter}
\alias{strwrap_signature_linter}
\title{Check signature of \code{\link[=strwrap]{strwrap()}}}
\usage{
strwrap_signature_linter()
}
\description{
Check signature of \code{\link[=strwrap]{strwrap()}}
}
