% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_signature_linter.R
\name{sub_signature_linter}
\alias{sub_signature_linter}
\title{Check signature of \code{\link[=sub]{sub()}}}
\usage{
sub_signature_linter()
}
\description{
Check signature of \code{\link[=sub]{sub()}}
}
