% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_matrix_signature_linter.R
\name{subset_matrix_signature_linter}
\alias{subset_matrix_signature_linter}
\title{Check signature of \code{\link[=subset.matrix]{subset.matrix()}}}
\usage{
subset_matrix_signature_linter()
}
\description{
Check signature of \code{\link[=subset.matrix]{subset.matrix()}}
}
