% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute_signature_linter.R
\name{substitute_signature_linter}
\alias{substitute_signature_linter}
\title{Check signature of \code{\link[=substitute]{substitute()}}}
\usage{
substitute_signature_linter()
}
\description{
Check signature of \code{\link[=substitute]{substitute()}}
}
