% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substr_signature_linter.R
\name{substr_signature_linter}
\alias{substr_signature_linter}
\title{Check signature of \code{\link[=substr]{substr()}}}
\usage{
substr_signature_linter()
}
\description{
Check signature of \code{\link[=substr]{substr()}}
}
