% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substring_signature_linter.R
\name{substring_signature_linter}
\alias{substring_signature_linter}
\title{Check signature of \code{\link[=substring]{substring()}}}
\usage{
substring_signature_linter()
}
\description{
Check signature of \code{\link[=substring]{substring()}}
}
