% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary_Date_signature_linter.R
\name{Summary_Date_signature_linter}
\alias{Summary_Date_signature_linter}
\title{Check signature of \code{\link[=Summary.Date]{Summary.Date()}}}
\usage{
Summary_Date_signature_linter()
}
\description{
Check signature of \code{\link[=Summary.Date]{Summary.Date()}}
}
