% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_factor_signature_linter.R
\name{summary_factor_signature_linter}
\alias{summary_factor_signature_linter}
\title{Check signature of \code{\link[=summary.factor]{summary.factor()}}}
\usage{
summary_factor_signature_linter()
}
\description{
Check signature of \code{\link[=summary.factor]{summary.factor()}}
}
