% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_proc_time_signature_linter.R
\name{summary_proc_time_signature_linter}
\alias{summary_proc_time_signature_linter}
\title{Check signature of \code{\link[=summary.proc_time]{summary.proc_time()}}}
\usage{
summary_proc_time_signature_linter()
}
\description{
Check signature of \code{\link[=summary.proc_time]{summary.proc_time()}}
}
