% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_srcfile_signature_linter.R
\name{summary_srcfile_signature_linter}
\alias{summary_srcfile_signature_linter}
\title{Check signature of \code{\link[=summary.srcfile]{summary.srcfile()}}}
\usage{
summary_srcfile_signature_linter()
}
\description{
Check signature of \code{\link[=summary.srcfile]{summary.srcfile()}}
}
