% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_srcref_signature_linter.R
\name{summary_srcref_signature_linter}
\alias{summary_srcref_signature_linter}
\title{Check signature of \code{\link[=summary.srcref]{summary.srcref()}}}
\usage{
summary_srcref_signature_linter()
}
\description{
Check signature of \code{\link[=summary.srcref]{summary.srcref()}}
}
