% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table_signature_linter.R
\name{summary_table_signature_linter}
\alias{summary_table_signature_linter}
\title{Check signature of \code{\link[=summary.table]{summary.table()}}}
\usage{
summary_table_signature_linter()
}
\description{
Check signature of \code{\link[=summary.table]{summary.table()}}
}
