% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_warnings_signature_linter.R
\name{summary_warnings_signature_linter}
\alias{summary_warnings_signature_linter}
\title{Check signature of \code{\link[=summary.warnings]{summary.warnings()}}}
\usage{
summary_warnings_signature_linter()
}
\description{
Check signature of \code{\link[=summary.warnings]{summary.warnings()}}
}
