% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppressMessages_signature_linter.R
\name{suppressMessages_signature_linter}
\alias{suppressMessages_signature_linter}
\title{Check signature of \code{\link[=suppressMessages]{suppressMessages()}}}
\usage{
suppressMessages_signature_linter()
}
\description{
Check signature of \code{\link[=suppressMessages]{suppressMessages()}}
}
