% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/suppressPackageStartupMessages_signature_linter.R
\name{suppressPackageStartupMessages_signature_linter}
\alias{suppressPackageStartupMessages_signature_linter}
\title{Check signature of \code{\link[=suppressPackageStartupMessages]{suppressPackageStartupMessages()}}}
\usage{
suppressPackageStartupMessages_signature_linter()
}
\description{
Check signature of \code{\link[=suppressPackageStartupMessages]{suppressPackageStartupMessages()}}
}
