% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppressWarnings_signature_linter.R
\name{suppressWarnings_signature_linter}
\alias{suppressWarnings_signature_linter}
\title{Check signature of \code{\link[=suppressWarnings]{suppressWarnings()}}}
\usage{
suppressWarnings_signature_linter()
}
\description{
Check signature of \code{\link[=suppressWarnings]{suppressWarnings()}}
}
