% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suspendInterrupts_signature_linter.R
\name{suspendInterrupts_signature_linter}
\alias{suspendInterrupts_signature_linter}
\title{Check signature of \code{\link[=suspendInterrupts]{suspendInterrupts()}}}
\usage{
suspendInterrupts_signature_linter()
}
\description{
Check signature of \code{\link[=suspendInterrupts]{suspendInterrupts()}}
}
