% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_signature_linter.R
\name{svd_signature_linter}
\alias{svd_signature_linter}
\title{Check signature of \code{\link[=svd]{svd()}}}
\usage{
svd_signature_linter()
}
\description{
Check signature of \code{\link[=svd]{svd()}}
}
