% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_call_signature_linter.R
\name{sys_call_signature_linter}
\alias{sys_call_signature_linter}
\title{Check signature of \code{\link[=sys.call]{sys.call()}}}
\usage{
sys_call_signature_linter()
}
\description{
Check signature of \code{\link[=sys.call]{sys.call()}}
}
