% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sys_chmod_signature_linter.R
\name{Sys_chmod_signature_linter}
\alias{Sys_chmod_signature_linter}
\title{Check signature of \code{\link[=Sys.chmod]{Sys.chmod()}}}
\usage{
Sys_chmod_signature_linter()
}
\description{
Check signature of \code{\link[=Sys.chmod]{Sys.chmod()}}
}
