% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frame_signature_linter.R
\name{sys_frame_signature_linter}
\alias{sys_frame_signature_linter}
\title{Check signature of \code{\link[=sys.frame]{sys.frame()}}}
\usage{
sys_frame_signature_linter()
}
\description{
Check signature of \code{\link[=sys.frame]{sys.frame()}}
}
