% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_function_signature_linter.R
\name{sys_function_signature_linter}
\alias{sys_function_signature_linter}
\title{Check signature of \code{\link[=sys.function]{sys.function()}}}
\usage{
sys_function_signature_linter()
}
\description{
Check signature of \code{\link[=sys.function]{sys.function()}}
}
