% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sys_getenv_signature_linter.R
\name{Sys_getenv_signature_linter}
\alias{Sys_getenv_signature_linter}
\title{Check signature of \code{\link[=Sys.getenv]{Sys.getenv()}}}
\usage{
Sys_getenv_signature_linter()
}
\description{
Check signature of \code{\link[=Sys.getenv]{Sys.getenv()}}
}
