% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sys_getlocale_signature_linter.R
\name{Sys_getlocale_signature_linter}
\alias{Sys_getlocale_signature_linter}
\title{Check signature of \code{\link[=Sys.getlocale]{Sys.getlocale()}}}
\usage{
Sys_getlocale_signature_linter()
}
\description{
Check signature of \code{\link[=Sys.getlocale]{Sys.getlocale()}}
}
