% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sys_glob_signature_linter.R
\name{Sys_glob_signature_linter}
\alias{Sys_glob_signature_linter}
\title{Check signature of \code{\link[=Sys.glob]{Sys.glob()}}}
\usage{
Sys_glob_signature_linter()
}
\description{
Check signature of \code{\link[=Sys.glob]{Sys.glob()}}
}
