% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_parent_signature_linter.R
\name{sys_parent_signature_linter}
\alias{sys_parent_signature_linter}
\title{Check signature of \code{\link[=sys.parent]{sys.parent()}}}
\usage{
sys_parent_signature_linter()
}
\description{
Check signature of \code{\link[=sys.parent]{sys.parent()}}
}
