% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sys_readlink_signature_linter.R
\name{Sys_readlink_signature_linter}
\alias{Sys_readlink_signature_linter}
\title{Check signature of \code{\link[=Sys.readlink]{Sys.readlink()}}}
\usage{
Sys_readlink_signature_linter()
}
\description{
Check signature of \code{\link[=Sys.readlink]{Sys.readlink()}}
}
