% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_save_image_signature_linter.R
\name{sys_save_image_signature_linter}
\alias{sys_save_image_signature_linter}
\title{Check signature of \code{\link[=sys.save.image]{sys.save.image()}}}
\usage{
sys_save_image_signature_linter()
}
\description{
Check signature of \code{\link[=sys.save.image]{sys.save.image()}}
}
