% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sys_setFileTime_signature_linter.R
\name{Sys_setFileTime_signature_linter}
\alias{Sys_setFileTime_signature_linter}
\title{Check signature of \code{\link[=Sys.setFileTime]{Sys.setFileTime()}}}
\usage{
Sys_setFileTime_signature_linter()
}
\description{
Check signature of \code{\link[=Sys.setFileTime]{Sys.setFileTime()}}
}
