% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sys_setLanguage_signature_linter.R
\name{Sys_setLanguage_signature_linter}
\alias{Sys_setLanguage_signature_linter}
\title{Check signature of \code{\link[=Sys.setLanguage]{Sys.setLanguage()}}}
\usage{
Sys_setLanguage_signature_linter()
}
\description{
Check signature of \code{\link[=Sys.setLanguage]{Sys.setLanguage()}}
}
