% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sys_setlocale_signature_linter.R
\name{Sys_setlocale_signature_linter}
\alias{Sys_setlocale_signature_linter}
\title{Check signature of \code{\link[=Sys.setlocale]{Sys.setlocale()}}}
\usage{
Sys_setlocale_signature_linter()
}
\description{
Check signature of \code{\link[=Sys.setlocale]{Sys.setlocale()}}
}
