% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sys_sleep_signature_linter.R
\name{Sys_sleep_signature_linter}
\alias{Sys_sleep_signature_linter}
\title{Check signature of \code{\link[=Sys.sleep]{Sys.sleep()}}}
\usage{
Sys_sleep_signature_linter()
}
\description{
Check signature of \code{\link[=Sys.sleep]{Sys.sleep()}}
}
