% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_source_signature_linter.R
\name{sys_source_signature_linter}
\alias{sys_source_signature_linter}
\title{Check signature of \code{\link[=sys.source]{sys.source()}}}
\usage{
sys_source_signature_linter()
}
\description{
Check signature of \code{\link[=sys.source]{sys.source()}}
}
