% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sys_timezone_signature_linter.R
\name{Sys_timezone_signature_linter}
\alias{Sys_timezone_signature_linter}
\title{Check signature of \code{\link[=Sys.timezone]{Sys.timezone()}}}
\usage{
Sys_timezone_signature_linter()
}
\description{
Check signature of \code{\link[=Sys.timezone]{Sys.timezone()}}
}
