% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sys_umask_signature_linter.R
\name{Sys_umask_signature_linter}
\alias{Sys_umask_signature_linter}
\title{Check signature of \code{\link[=Sys.umask]{Sys.umask()}}}
\usage{
Sys_umask_signature_linter()
}
\description{
Check signature of \code{\link[=Sys.umask]{Sys.umask()}}
}
