% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sys_which_signature_linter.R
\name{Sys_which_signature_linter}
\alias{Sys_which_signature_linter}
\title{Check signature of \code{\link[=Sys.which]{Sys.which()}}}
\usage{
Sys_which_signature_linter()
}
\description{
Check signature of \code{\link[=Sys.which]{Sys.which()}}
}
