% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_signature_linter.R
\name{system_signature_linter}
\alias{system_signature_linter}
\title{Check signature of \code{\link[=system]{system()}}}
\usage{
system_signature_linter()
}
\description{
Check signature of \code{\link[=system]{system()}}
}
