% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_time_signature_linter.R
\name{system_time_signature_linter}
\alias{system_time_signature_linter}
\title{Check signature of \code{\link[=system.time]{system.time()}}}
\usage{
system_time_signature_linter()
}
\description{
Check signature of \code{\link[=system.time]{system.time()}}
}
