% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_data_frame_signature_linter.R
\name{t_data_frame_signature_linter}
\alias{t_data_frame_signature_linter}
\title{Check signature of \code{\link[=t.data.frame]{t.data.frame()}}}
\usage{
t_data_frame_signature_linter()
}
\description{
Check signature of \code{\link[=t.data.frame]{t.data.frame()}}
}
