% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_default_signature_linter.R
\name{t_default_signature_linter}
\alias{t_default_signature_linter}
\title{Check signature of \code{\link[=t.default]{t.default()}}}
\usage{
t_default_signature_linter()
}
\description{
Check signature of \code{\link[=t.default]{t.default()}}
}
