% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_signature_linter.R
\name{tabulate_signature_linter}
\alias{tabulate_signature_linter}
\title{Check signature of \code{\link[=tabulate]{tabulate()}}}
\usage{
tabulate_signature_linter()
}
\description{
Check signature of \code{\link[=tabulate]{tabulate()}}
}
