% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tan_signature_linter.R
\name{tan_signature_linter}
\alias{tan_signature_linter}
\title{Check signature of \code{\link[=tan]{tan()}}}
\usage{
tan_signature_linter()
}
\description{
Check signature of \code{\link[=tan]{tan()}}
}
