% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanh_signature_linter.R
\name{tanh_signature_linter}
\alias{tanh_signature_linter}
\title{Check signature of \code{\link[=tanh]{tanh()}}}
\usage{
tanh_signature_linter()
}
\description{
Check signature of \code{\link[=tanh]{tanh()}}
}
