% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanpi_signature_linter.R
\name{tanpi_signature_linter}
\alias{tanpi_signature_linter}
\title{Check signature of \code{\link[=tanpi]{tanpi()}}}
\usage{
tanpi_signature_linter()
}
\description{
Check signature of \code{\link[=tanpi]{tanpi()}}
}
