% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskCallbackManager_signature_linter.R
\name{taskCallbackManager_signature_linter}
\alias{taskCallbackManager_signature_linter}
\title{Check signature of \code{\link[=taskCallbackManager]{taskCallbackManager()}}}
\usage{
taskCallbackManager_signature_linter()
}
\description{
Check signature of \code{\link[=taskCallbackManager]{taskCallbackManager()}}
}
