% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdir_signature_linter.R
\name{tempdir_signature_linter}
\alias{tempdir_signature_linter}
\title{Check signature of \code{\link[=tempdir]{tempdir()}}}
\usage{
tempdir_signature_linter()
}
\description{
Check signature of \code{\link[=tempdir]{tempdir()}}
}
