% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempfile_signature_linter.R
\name{tempfile_signature_linter}
\alias{tempfile_signature_linter}
\title{Check signature of \code{\link[=tempfile]{tempfile()}}}
\usage{
tempfile_signature_linter()
}
\description{
Check signature of \code{\link[=tempfile]{tempfile()}}
}
