% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textConnection_signature_linter.R
\name{textConnection_signature_linter}
\alias{textConnection_signature_linter}
\title{Check signature of \code{\link[=textConnection]{textConnection()}}}
\usage{
textConnection_signature_linter()
}
\description{
Check signature of \code{\link[=textConnection]{textConnection()}}
}
