% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textConnectionValue_signature_linter.R
\name{textConnectionValue_signature_linter}
\alias{textConnectionValue_signature_linter}
\title{Check signature of \code{\link[=textConnectionValue]{textConnectionValue()}}}
\usage{
textConnectionValue_signature_linter()
}
\description{
Check signature of \code{\link[=textConnectionValue]{textConnectionValue()}}
}
