% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolower_signature_linter.R
\name{tolower_signature_linter}
\alias{tolower_signature_linter}
\title{Check signature of \code{\link[=tolower]{tolower()}}}
\usage{
tolower_signature_linter()
}
\description{
Check signature of \code{\link[=tolower]{tolower()}}
}
