% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topenv_signature_linter.R
\name{topenv_signature_linter}
\alias{topenv_signature_linter}
\title{Check signature of \code{\link[=topenv]{topenv()}}}
\usage{
topenv_signature_linter()
}
\description{
Check signature of \code{\link[=topenv]{topenv()}}
}
