% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toupper_signature_linter.R
\name{toupper_signature_linter}
\alias{toupper_signature_linter}
\title{Check signature of \code{\link[=toupper]{toupper()}}}
\usage{
toupper_signature_linter()
}
\description{
Check signature of \code{\link[=toupper]{toupper()}}
}
