% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_signature_linter.R
\name{trace_signature_linter}
\alias{trace_signature_linter}
\title{Check signature of \code{\link[=trace]{trace()}}}
\usage{
trace_signature_linter()
}
\description{
Check signature of \code{\link[=trace]{trace()}}
}
