% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceback_signature_linter.R
\name{traceback_signature_linter}
\alias{traceback_signature_linter}
\title{Check signature of \code{\link[=traceback]{traceback()}}}
\usage{
traceback_signature_linter()
}
\description{
Check signature of \code{\link[=traceback]{traceback()}}
}
