% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracingState_signature_linter.R
\name{tracingState_signature_linter}
\alias{tracingState_signature_linter}
\title{Check signature of \code{\link[=tracingState]{tracingState()}}}
\usage{
tracingState_signature_linter()
}
\description{
Check signature of \code{\link[=tracingState]{tracingState()}}
}
