% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigamma_signature_linter.R
\name{trigamma_signature_linter}
\alias{trigamma_signature_linter}
\title{Check signature of \code{\link[=trigamma]{trigamma()}}}
\usage{
trigamma_signature_linter()
}
\description{
Check signature of \code{\link[=trigamma]{trigamma()}}
}
