% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimws_signature_linter.R
\name{trimws_signature_linter}
\alias{trimws_signature_linter}
\title{Check signature of \code{\link[=trimws]{trimws()}}}
\usage{
trimws_signature_linter()
}
\description{
Check signature of \code{\link[=trimws]{trimws()}}
}
