% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_Date_signature_linter.R
\name{trunc_Date_signature_linter}
\alias{trunc_Date_signature_linter}
\title{Check signature of \code{\link[=trunc.Date]{trunc.Date()}}}
\usage{
trunc_Date_signature_linter()
}
\description{
Check signature of \code{\link[=trunc.Date]{trunc.Date()}}
}
