% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_POSIXt_signature_linter.R
\name{trunc_POSIXt_signature_linter}
\alias{trunc_POSIXt_signature_linter}
\title{Check signature of \code{\link[=trunc.POSIXt]{trunc.POSIXt()}}}
\usage{
trunc_POSIXt_signature_linter()
}
\description{
Check signature of \code{\link[=trunc.POSIXt]{trunc.POSIXt()}}
}
