% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_signature_linter.R
\name{try_signature_linter}
\alias{try_signature_linter}
\title{Check signature of \code{\link[=try]{try()}}}
\usage{
try_signature_linter()
}
\description{
Check signature of \code{\link[=try]{try()}}
}
