% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeof_signature_linter.R
\name{typeof_signature_linter}
\alias{typeof_signature_linter}
\title{Check signature of \code{\link[=typeof]{typeof()}}}
\usage{
typeof_signature_linter()
}
\description{
Check signature of \code{\link[=typeof]{typeof()}}
}
