% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unclass_signature_linter.R
\name{unclass_signature_linter}
\alias{unclass_signature_linter}
\title{Check signature of \code{\link[=unclass]{unclass()}}}
\usage{
unclass_signature_linter()
}
\description{
Check signature of \code{\link[=unclass]{unclass()}}
}
