% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undebug_signature_linter.R
\name{undebug_signature_linter}
\alias{undebug_signature_linter}
\title{Check signature of \code{\link[=undebug]{undebug()}}}
\usage{
undebug_signature_linter()
}
\description{
Check signature of \code{\link[=undebug]{undebug()}}
}
