% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union_signature_linter.R
\name{union_signature_linter}
\alias{union_signature_linter}
\title{Check signature of \code{\link[=union]{union()}}}
\usage{
union_signature_linter()
}
\description{
Check signature of \code{\link[=union]{union()}}
}
