% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_signature_linter.R
\name{units_signature_linter}
\alias{units_signature_linter}
\title{Check signature of \code{\link[=units]{units()}}}
\usage{
units_signature_linter()
}
\description{
Check signature of \code{\link[=units]{units()}}
}
