% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlink_signature_linter.R
\name{unlink_signature_linter}
\alias{unlink_signature_linter}
\title{Check signature of \code{\link[=unlink]{unlink()}}}
\usage{
unlink_signature_linter()
}
\description{
Check signature of \code{\link[=unlink]{unlink()}}
}
