% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_signature_linter.R
\name{unlist_signature_linter}
\alias{unlist_signature_linter}
\title{Check signature of \code{\link[=unlist]{unlist()}}}
\usage{
unlist_signature_linter()
}
\description{
Check signature of \code{\link[=unlist]{unlist()}}
}
