% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unloadNamespace_signature_linter.R
\name{unloadNamespace_signature_linter}
\alias{unloadNamespace_signature_linter}
\title{Check signature of \code{\link[=unloadNamespace]{unloadNamespace()}}}
\usage{
unloadNamespace_signature_linter()
}
\description{
Check signature of \code{\link[=unloadNamespace]{unloadNamespace()}}
}
