% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlockBinding_signature_linter.R
\name{unlockBinding_signature_linter}
\alias{unlockBinding_signature_linter}
\title{Check signature of \code{\link[=unlockBinding]{unlockBinding()}}}
\usage{
unlockBinding_signature_linter()
}
\description{
Check signature of \code{\link[=unlockBinding]{unlockBinding()}}
}
