% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unname_signature_linter.R
\name{unname_signature_linter}
\alias{unname_signature_linter}
\title{Check signature of \code{\link[=unname]{unname()}}}
\usage{
unname_signature_linter()
}
\description{
Check signature of \code{\link[=unname]{unname()}}
}
