% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unserialize_signature_linter.R
\name{unserialize_signature_linter}
\alias{unserialize_signature_linter}
\title{Check signature of \code{\link[=unserialize]{unserialize()}}}
\usage{
unserialize_signature_linter()
}
\description{
Check signature of \code{\link[=unserialize]{unserialize()}}
}
