% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsplit_signature_linter.R
\name{unsplit_signature_linter}
\alias{unsplit_signature_linter}
\title{Check signature of \code{\link[=unsplit]{unsplit()}}}
\usage{
unsplit_signature_linter()
}
\description{
Check signature of \code{\link[=unsplit]{unsplit()}}
}
