% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untrace_signature_linter.R
\name{untrace_signature_linter}
\alias{untrace_signature_linter}
\title{Check signature of \code{\link[=untrace]{untrace()}}}
\usage{
untrace_signature_linter()
}
\description{
Check signature of \code{\link[=untrace]{untrace()}}
}
