% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untracemem_signature_linter.R
\name{untracemem_signature_linter}
\alias{untracemem_signature_linter}
\title{Check signature of \code{\link[=untracemem]{untracemem()}}}
\usage{
untracemem_signature_linter()
}
\description{
Check signature of \code{\link[=untracemem]{untracemem()}}
}
