% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unz_signature_linter.R
\name{unz_signature_linter}
\alias{unz_signature_linter}
\title{Check signature of \code{\link[=unz]{unz()}}}
\usage{
unz_signature_linter()
}
\description{
Check signature of \code{\link[=unz]{unz()}}
}
