% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper_tri_signature_linter.R
\name{upper_tri_signature_linter}
\alias{upper_tri_signature_linter}
\title{Check signature of \code{\link[=upper.tri]{upper.tri()}}}
\usage{
upper_tri_signature_linter()
}
\description{
Check signature of \code{\link[=upper.tri]{upper.tri()}}
}
