% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_signature_linter.R
\name{use_signature_linter}
\alias{use_signature_linter}
\title{Check signature of \code{\link[=use]{use()}}}
\usage{
use_signature_linter()
}
\description{
Check signature of \code{\link[=use]{use()}}
}
