% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UseMethod_signature_linter.R
\name{UseMethod_signature_linter}
\alias{UseMethod_signature_linter}
\title{Check signature of \code{\link[=UseMethod]{UseMethod()}}}
\usage{
UseMethod_signature_linter()
}
\description{
Check signature of \code{\link[=UseMethod]{UseMethod()}}
}
