% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validEnc_signature_linter.R
\name{validEnc_signature_linter}
\alias{validEnc_signature_linter}
\title{Check signature of \code{\link[=validEnc]{validEnc()}}}
\usage{
validEnc_signature_linter()
}
\description{
Check signature of \code{\link[=validEnc]{validEnc()}}
}
