% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_signature_linter.R
\name{vector_signature_linter}
\alias{vector_signature_linter}
\title{Check signature of \code{\link[=vector]{vector()}}}
\usage{
vector_signature_linter()
}
\description{
Check signature of \code{\link[=vector]{vector()}}
}
