% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vectorize_signature_linter.R
\name{Vectorize_signature_linter}
\alias{Vectorize_signature_linter}
\title{Check signature of \code{\link[=Vectorize]{Vectorize()}}}
\usage{
Vectorize_signature_linter()
}
\description{
Check signature of \code{\link[=Vectorize]{Vectorize()}}
}
