% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekdays_Date_signature_linter.R
\name{weekdays_Date_signature_linter}
\alias{weekdays_Date_signature_linter}
\title{Check signature of \code{\link[=weekdays.Date]{weekdays.Date()}}}
\usage{
weekdays_Date_signature_linter()
}
\description{
Check signature of \code{\link[=weekdays.Date]{weekdays.Date()}}
}
