% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekdays_POSIXt_signature_linter.R
\name{weekdays_POSIXt_signature_linter}
\alias{weekdays_POSIXt_signature_linter}
\title{Check signature of \code{\link[=weekdays.POSIXt]{weekdays.POSIXt()}}}
\usage{
weekdays_POSIXt_signature_linter()
}
\description{
Check signature of \code{\link[=weekdays.POSIXt]{weekdays.POSIXt()}}
}
