% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_max_signature_linter.R
\name{which_max_signature_linter}
\alias{which_max_signature_linter}
\title{Check signature of \code{\link[=which.max]{which.max()}}}
\usage{
which_max_signature_linter()
}
\description{
Check signature of \code{\link[=which.max]{which.max()}}
}
