% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_min_signature_linter.R
\name{which_min_signature_linter}
\alias{which_min_signature_linter}
\title{Check signature of \code{\link[=which.min]{which.min()}}}
\usage{
which_min_signature_linter()
}
\description{
Check signature of \code{\link[=which.min]{which.min()}}
}
