% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_signature_linter.R
\name{which_signature_linter}
\alias{which_signature_linter}
\title{Check signature of \code{\link[=which]{which()}}}
\usage{
which_signature_linter()
}
\description{
Check signature of \code{\link[=which]{which()}}
}
