% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withAutoprint_signature_linter.R
\name{withAutoprint_signature_linter}
\alias{withAutoprint_signature_linter}
\title{Check signature of \code{\link[=withAutoprint]{withAutoprint()}}}
\usage{
withAutoprint_signature_linter()
}
\description{
Check signature of \code{\link[=withAutoprint]{withAutoprint()}}
}
