% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withVisible_signature_linter.R
\name{withVisible_signature_linter}
\alias{withVisible_signature_linter}
\title{Check signature of \code{\link[=withVisible]{withVisible()}}}
\usage{
withVisible_signature_linter()
}
\description{
Check signature of \code{\link[=withVisible]{withVisible()}}
}
