% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_signature_linter.R
\name{write_signature_linter}
\alias{write_signature_linter}
\title{Check signature of \code{\link[=write]{write()}}}
\usage{
write_signature_linter()
}
\description{
Check signature of \code{\link[=write]{write()}}
}
