% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeBin_signature_linter.R
\name{writeBin_signature_linter}
\alias{writeBin_signature_linter}
\title{Check signature of \code{\link[=writeBin]{writeBin()}}}
\usage{
writeBin_signature_linter()
}
\description{
Check signature of \code{\link[=writeBin]{writeBin()}}
}
