% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeChar_signature_linter.R
\name{writeChar_signature_linter}
\alias{writeChar_signature_linter}
\title{Check signature of \code{\link[=writeChar]{writeChar()}}}
\usage{
writeChar_signature_linter()
}
\description{
Check signature of \code{\link[=writeChar]{writeChar()}}
}
