% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLines_signature_linter.R
\name{writeLines_signature_linter}
\alias{writeLines_signature_linter}
\title{Check signature of \code{\link[=writeLines]{writeLines()}}}
\usage{
writeLines_signature_linter()
}
\description{
Check signature of \code{\link[=writeLines]{writeLines()}}
}
