% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xor_signature_linter.R
\name{xor_signature_linter}
\alias{xor_signature_linter}
\title{Check signature of \code{\link[=xor]{xor()}}}
\usage{
xor_signature_linter()
}
\description{
Check signature of \code{\link[=xor]{xor()}}
}
