% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdrows_data_frame_signature_linter.R
\name{xpdrows_data_frame_signature_linter}
\alias{xpdrows_data_frame_signature_linter}
\title{Check signature of \code{\link[=xpdrows.data.frame]{xpdrows.data.frame()}}}
\usage{
xpdrows_data_frame_signature_linter()
}
\description{
Check signature of \code{\link[=xpdrows.data.frame]{xpdrows.data.frame()}}
}
