% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtfrm_AsIs_signature_linter.R
\name{xtfrm_AsIs_signature_linter}
\alias{xtfrm_AsIs_signature_linter}
\title{Check signature of \code{\link[=xtfrm.AsIs]{xtfrm.AsIs()}}}
\usage{
xtfrm_AsIs_signature_linter()
}
\description{
Check signature of \code{\link[=xtfrm.AsIs]{xtfrm.AsIs()}}
}
