% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtfrm_data_frame_signature_linter.R
\name{xtfrm_data_frame_signature_linter}
\alias{xtfrm_data_frame_signature_linter}
\title{Check signature of \code{\link[=xtfrm.data.frame]{xtfrm.data.frame()}}}
\usage{
xtfrm_data_frame_signature_linter()
}
\description{
Check signature of \code{\link[=xtfrm.data.frame]{xtfrm.data.frame()}}
}
