% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtfrm_Date_signature_linter.R
\name{xtfrm_Date_signature_linter}
\alias{xtfrm_Date_signature_linter}
\title{Check signature of \code{\link[=xtfrm.Date]{xtfrm.Date()}}}
\usage{
xtfrm_Date_signature_linter()
}
\description{
Check signature of \code{\link[=xtfrm.Date]{xtfrm.Date()}}
}
