% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtfrm_default_signature_linter.R
\name{xtfrm_default_signature_linter}
\alias{xtfrm_default_signature_linter}
\title{Check signature of \code{\link[=xtfrm.default]{xtfrm.default()}}}
\usage{
xtfrm_default_signature_linter()
}
\description{
Check signature of \code{\link[=xtfrm.default]{xtfrm.default()}}
}
