% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtfrm_difftime_signature_linter.R
\name{xtfrm_difftime_signature_linter}
\alias{xtfrm_difftime_signature_linter}
\title{Check signature of \code{\link[=xtfrm.difftime]{xtfrm.difftime()}}}
\usage{
xtfrm_difftime_signature_linter()
}
\description{
Check signature of \code{\link[=xtfrm.difftime]{xtfrm.difftime()}}
}
