% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtfrm_factor_signature_linter.R
\name{xtfrm_factor_signature_linter}
\alias{xtfrm_factor_signature_linter}
\title{Check signature of \code{\link[=xtfrm.factor]{xtfrm.factor()}}}
\usage{
xtfrm_factor_signature_linter()
}
\description{
Check signature of \code{\link[=xtfrm.factor]{xtfrm.factor()}}
}
