% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtfrm_numeric_version_signature_linter.R
\name{xtfrm_numeric_version_signature_linter}
\alias{xtfrm_numeric_version_signature_linter}
\title{Check signature of \code{\link[=xtfrm.numeric_version]{xtfrm.numeric_version()}}}
\usage{
xtfrm_numeric_version_signature_linter()
}
\description{
Check signature of \code{\link[=xtfrm.numeric_version]{xtfrm.numeric_version()}}
}
