% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtfrm_POSIXct_signature_linter.R
\name{xtfrm_POSIXct_signature_linter}
\alias{xtfrm_POSIXct_signature_linter}
\title{Check signature of \code{\link[=xtfrm.POSIXct]{xtfrm.POSIXct()}}}
\usage{
xtfrm_POSIXct_signature_linter()
}
\description{
Check signature of \code{\link[=xtfrm.POSIXct]{xtfrm.POSIXct()}}
}
