% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtfrm_POSIXlt_signature_linter.R
\name{xtfrm_POSIXlt_signature_linter}
\alias{xtfrm_POSIXlt_signature_linter}
\title{Check signature of \code{\link[=xtfrm.POSIXlt]{xtfrm.POSIXlt()}}}
\usage{
xtfrm_POSIXlt_signature_linter()
}
\description{
Check signature of \code{\link[=xtfrm.POSIXlt]{xtfrm.POSIXlt()}}
}
