% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xzfile_signature_linter.R
\name{xzfile_signature_linter}
\alias{xzfile_signature_linter}
\title{Check signature of \code{\link[=xzfile]{xzfile()}}}
\usage{
xzfile_signature_linter()
}
\description{
Check signature of \code{\link[=xzfile]{xzfile()}}
}
