% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zapsmall_signature_linter.R
\name{zapsmall_signature_linter}
\alias{zapsmall_signature_linter}
\title{Check signature of \code{\link[=zapsmall]{zapsmall()}}}
\usage{
zapsmall_signature_linter()
}
\description{
Check signature of \code{\link[=zapsmall]{zapsmall()}}
}
