% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zstdfile_signature_linter.R
\name{zstdfile_signature_linter}
\alias{zstdfile_signature_linter}
\title{Check signature of \code{\link[=zstdfile]{zstdfile()}}}
\usage{
zstdfile_signature_linter()
}
\description{
Check signature of \code{\link[=zstdfile]{zstdfile()}}
}
