% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCOL_signature_linter.R
\name{NCOL_signature_linter}
\alias{NCOL_signature_linter}
\title{Check signature of \code{\link[=NCOL]{NCOL()}}}
\usage{
NCOL_signature_linter()
}
\description{
Check signature of \code{\link[=NCOL]{NCOL()}}
}
