% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrow_signature_linter.R
\name{nrow_signature_linter}
\alias{nrow_signature_linter}
\title{Check signature of \code{\link[=nrow]{nrow()}}}
\usage{
nrow_signature_linter()
}
\description{
Check signature of \code{\link[=nrow]{nrow()}}
}
