% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary_factor_signature_linter.R
\name{Summary_factor_signature_linter}
\alias{Summary_factor_signature_linter}
\title{Check signature of \code{\link[=Summary.factor]{Summary.factor()}}}
\usage{
Summary_factor_signature_linter()
}
\description{
Check signature of \code{\link[=Summary.factor]{Summary.factor()}}
}
